% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_ui.R
\name{AddAggsAdjacencyMatrix}
\alias{AddAggsAdjacencyMatrix}
\title{Construct tile adjacency matrix from consolidated GetTiles output.}
\usage{
AddAggsAdjacencyMatrix(aggs)
}
\arguments{
\item{aggs}{Aggregated tile information after consolidation.}
}
\description{
Construct tile adjacency matrix from consolidated GetTiles output.
}
