% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_analysis.R
\name{AddSubClusterLabels}
\alias{AddSubClusterLabels}
\title{Add sub-cluster labels to a Seurat object}
\usage{
AddSubClusterLabels(
  obj,
  obj_sub,
  cluster,
  clusters.name = "seurat_clusters",
  sub.clusters.name = NULL
)
}
\arguments{
\item{obj}{A Seurat object.}

\item{obj_sub}{A Seurat object containing the sub-clustered cells.}

\item{cluster}{The parent cluster label in \code{obj} that was sub-clustered.}

\item{clusters.name}{Name of the metadata column in \code{obj} containing the parent cluster labels.}

\item{sub.clusters.name}{Name of the metadata column to store the sub-cluster labels.
If NULL, defaults to \verb{<clusters.name>_s<cluster>}.}
}
\value{
The input Seurat object with sub-cluster labels added to metadata.
}
\description{
Add sub-cluster labels to a Seurat object
}
