% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_analysis.R
\name{AddTileMetadata}
\alias{AddTileMetadata}
\title{Add tile-level metadata to a matching cell-level Seurat object}
\usage{
AddTileMetadata(obj, tile_obj, ..., by = c(tile_id = "id"))
}
\arguments{
\item{obj}{A Seurat object containing cell-level data with a \code{tile_id} metadata column.}

\item{tile_obj}{A Seurat object containing tile-level data.}

\item{...}{Tidy expressions specifying which metadata columns from \code{tile_obj}
to add to \code{obj}. For example, \code{tile_size = npts}}

\item{by}{A named character vector specifying how to join \code{obj} and \code{tile_obj}.
Defaults to \code{c('tile_id' = 'id')}, where \code{tile_id} is the metadata column
in \code{obj} and \code{id} is the metadata column in \code{tile_obj}.}
}
\value{
The input Seurat object with additional tile-level metadata added to \verb{@meta.data}.
}
\description{
Add tile-level metadata to a matching cell-level Seurat object
}
