% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{calc_moran}
\alias{calc_moran}
\title{Fast computation of Moran's I.}
\usage{
calc_moran(x, c1, c2)
}
\arguments{
\item{x}{Measurements.}

\item{c1}{x-coordinates.}

\item{c2}{y-coordinates.}
}
\value{
I Moran's I.

EI Expected value of Moran's I under the null of no
spatial correlation.

SD Standard deviation of Moran's I.
}
\description{
Fast computation of Moran's I.
}
\author{
Matthew Cooper.
}
