% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/post_analysis.R
\name{calc_Wald_pvalue_from_fdrtool}
\alias{calc_Wald_pvalue_from_fdrtool}
\title{Given Wald statistics, compute p-values.}
\usage{
calc_Wald_pvalue_from_fdrtool(wald_stats, params)
}
\arguments{
\item{wald_stats}{A numeric vector of Wald statistics.
Note: This function expects signed t-statistics, or, the un-squared statistics.}

\item{params}{The dataframe of parameters from \code{fit_empirical_null} or from
\code{fdrtool::fdrtool}; has one row and a column called \code{sd}.}
}
\value{
Vector of p-values (un-adjusted for multiple corrections).
}
\description{
In the TESSERA algorithm, and in generalized linear mixed models in general,
the exact distribution of p-values for Wald statistics under the null hypothesis
is unknown.
We know that the standard errors are, for finite samples, likely biased
downward (too small).
Moreover, for finite samples, due to the approximations
in the E-step of the fitting algorithm, there may be a small positive bias.
Hence, we use the fdrtool fndr method, which assumes that the null is a
centered normal distribution with some variance (estimated).
This function assumes that \code{fit_empirical_null} has been run; that function also
provides p-values, but, if you want to fit a distribution on one set of statistics
and then evaluate p-values on another set of statistics using the fitted distribution,
this function can be used.
}
\examples{
# Simulate null and non-null Wald statistics
null_stats <- rnorm(n = 2000, mean = 0.1, sd = 4)
alt_stats <- rnorm(n = 500, mean = 5, sd = 2)
wald_vec <- c(null_stats, alt_stats)

# Fit
emp_null <- fit_empirical_null_fdrtool(wald_vec)
calc_Wald_pvalue_from_fdrtool(wald_vec, emp_null$params)
}
\author{
Florica J Constantine, florica AT berkeley.edu
}
