% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/comparisons.R
\name{CARBayes_Leroux_wrapper}
\alias{CARBayes_Leroux_wrapper}
\title{Wrapper to fit a Leroux model with MCMC (CARBayes).}
\usage{
CARBayes_Leroux_wrapper(
  z_list,
  X_list,
  W_list,
  model_family = "poisson",
  z_offset = 0.5,
  chains = 4,
  iter = 2000,
  warmup = 200,
  thin = 2,
  cores = 1,
  library_size_list = NULL
)
}
\arguments{
\item{z_list}{List of vectors of observed counts---one vector per area.}

\item{X_list}{List of design or covariate matrices---one matrix per area.
Same length and ordering as z_list.
Matrices with number of rows equal to length of corresponding vector in
z_list.}

\item{W_list}{List of spatial adjacency matrices---one matrix per area.}

\item{model_family}{"poisson" or "gaussian".
Which model family to fit. Note that Poisson models don't scale well in the
multi-area setting---the spline is fit on a per-area basis.}

\item{z_offset}{In the "gaussian" case, the counts z are transformed as
log(z + z_offset).}

\item{chains}{Number of MCMC chains to run.}

\item{iter}{Number of total iterations per MCMC chain.}

\item{warmup}{Burn-in for MCMC.}

\item{thin}{Thinning rate for MCMC.}

\item{cores}{Number of cores to use for MCMC.}

\item{library_size_list}{A list with a vector of library sizes for each sample.
Total counts for each location if there are more than one gene/measurement;
if NULL, set to 1.}
}
\value{
The output list from CARBayes::S.CARleroux, plus a time field
for how long the function ran for.
}
\description{
This function theoretically works for multiple areas by creating a
block diagonal adjacency matrix. However, the CARBayes package
needs a dense, matrix-type object, so this WILL NOT SCALE!
}
\note{
Requires the Matrix library.

Requires the CARBayes library.
}
\author{
Florica J Constantine, florica AT berkeley.edu
}
