% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TESSERA_spNNGP.R
\name{checkInputsTESSERAspNNGP}
\alias{checkInputsTESSERAspNNGP}
\title{Check inputs for the TESSERA_spNNGP method.
Make sure that the input object has everything needed to run without error.}
\usage{
checkInputsTESSERAspNNGP(TESSERAData_obj)
}
\arguments{
\item{TESSERAData_obj}{Object containing data.
Created by the prepData method.}
}
\value{
Nothing.
}
\description{
Check inputs for the TESSERA_spNNGP method.
Make sure that the input object has everything needed to run without error.
}
\note{
Does not return anything.

This method can be used to check a hand-created input object.
E.g., if a user does not want to use prepData.
}
\examples{
# Locate the prepped TESSERAData object in inst/extdata
rds_path <- system.file("extdata", "example_prepData.rds", package = "TESSERA")
# Load the TESSERAData object
TESSERA_data <- readRDS(rds_path)

# Validate inputs for the spNNGP-based TESSERA model
checkInputsTESSERAspNNGP(TESSERA_data)
}
\author{
Florica J Constantine, florica AT berkeley.edu
}
