% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{compress_field_cpp}
\alias{compress_field_cpp}
\title{Compress a gradient field using SVD}
\usage{
compress_field_cpp(field)
}
\arguments{
\item{field}{A \code{2} x \code{D} x \code{N} array in column-major ordering
containing the spatial gradient in expression for each of
\code{D} latent variables at every point, edge, or triangle.}
}
\value{
A \code{N} x \code{6} matrix with the following attributes for
each location:
\item{dx grad,dy grad}{x,y directions of unit vector in the
direction of greatest change (first singular vector).}
\item{dx ortho,dy ortho}{x,y directions of unit vector orthogonal
to the direction of greatest change (second singular vector).}
\item{|grad|,|ortho|}{Magnitude of directional derivative in the
gradient and orthogonal directions (singular values).}
}
\description{
Expresses the \code{2} x \code{D} total derivative at each location as
a pair of \code{2}-dimensional vectors in the gradient and orthogonal
directions.
}
