% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_gradients.R
\name{compress_gradients_svd}
\alias{compress_gradients_svd}
\title{Compress a gradient field using SVD}
\usage{
compress_gradients_svd(field)
}
\arguments{
\item{field}{A gradient field with the following attributes:
\itemize{
\item \code{pts}: A \code{2} x \code{D} x \code{N} array in column-major ordering
containing the spatial gradient in expression for each of
\code{D} latent variables at every point in space.
\item \code{tris}: A \code{2} x \code{D} x \code{F} array in column-major ordering
containing the spatial gradient in expression for each of
\code{D} latent variables at every triangle in the mesh.
Average of the vertices (3 for full triangles, 2 for degenerate triangles).
\item \code{edges_pts}: A \code{2} x \code{D} x \code{E} array in column-major ordering
containing the spatial gradient in expression for each of
\code{D} latent variables at every primal edge (point-to-point) in the mesh.
Sum of the two endpoints.
\item \code{edges_tris}: A \code{2} x \code{D} x \code{E} array in column-major ordering
containing the spatial gradient in expression for each of
\code{D} latent variables at every dual edge (triangle-to-triangle) in the mesh.
Sum of the two adjacent triangles.
}}
}
\value{
A gradient field with the same attributes as the input, as well as compressed
representations \code{pts_svd}, \code{tris_svd}, \code{edges_pts_svd}, and \code{edges_tris_svd}. Each of these
is a \code{N} x \code{6} matrix with the following columns for
each location:
\item{dx_grad,dy_grad}{x,y directions of unit vector in the
direction of greatest change (first singular vector).}
\item{dx_ortho,dy_ortho}{x,y directions of unit vector orthogonal
to the direction of greatest change (second singular vector).}
\item{len_grad,len_ortho}{Magnitude of directional derivative in the
gradient and orthogonal directions (singular values).}
}
\description{
Expresses the \code{2} x \code{D} total derivative at each location as
a pair of \code{2}-dimensional vectors in the gradient and orthogonal
directions.
}
