% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_gradients.R
\name{compute_gradients}
\alias{compute_gradients}
\title{Compute spatial gradient field for input to DMT}
\usage{
compute_gradients(
  dmt,
  smooth_distance = "none",
  smooth_similarity = "none",
  smooth_iter = 1,
  on_edges = FALSE,
  edge_from_tri = FALSE,
  edge_from_pt = FALSE
)
}
\arguments{
\item{dmt}{A list containing the mesh data structures:
\itemize{
\item \code{pts} is a \code{N} x \code{2+M} data table with columns \code{X} and \code{Y} containing the
coordinates of cells and additional metadata.
\item \code{tris} is a \code{F} x \code{4} data table containing the X,Y coordinates of each
triangle's centroid in the first two columns, and area and
largest height of each triangle in the last two columns.
\item \code{edges} is a \code{E} x \code{14} data table with columns \code{from_pt}, \code{to_pt}, \code{from_tri}, \code{to_tri},
\code{x0_pt}, \code{x1_pt}, \code{y0_pt}, \code{y1_pt}, \code{x0_tri}, \code{x1_tri}, \code{y0_tri}, \code{y1_tri},
\code{length_pt}, \code{length_tri}. If only one triangle uses an edge, then the \code{from_tri},
\code{x0_tri}, and \code{y0_tri} fields will contain NaN values.
\item \code{tri_to_pt} is a \code{F} x \code{N} sparse matrix with value 1 at (i,j) if
triangle i uses point j as a vertex.
\item \code{udv_cells} contains cell embeddings stored in \code{embeddings}
(a \code{N} x \code{D} matrix with \code{D}-dimensional embeddings for each cell)
and \code{loadings} (a \code{G} x \code{D} matrix with gene loadings for each latent variable).
}}

\item{smooth_distance}{One of \code{c('none', 'euclidean', 'projected', 'constant')}.
If either \code{smooth_distance} or \code{smooth_similarity} is \code{'none'} (the default),
then no smoothing of the gradient field is conducted.}

\item{smooth_similarity}{One of \code{c('none', 'euclidean', 'projected', 'constant')}.
If either \code{smooth_distance} or \code{smooth_similarity} is \code{'none'} (the default),
then no smoothing of the gradient field is conducted.}

\item{smooth_iter}{Number of rounds of gradient smoothing.}
}
\value{
A gradient field with the following attributes:
\item{pts}{A \code{2} x \code{D} x \code{N} array in column-major ordering
containing the spatial gradient in expression for each of
\code{D} latent variables at every point in space.}
\item{tris}{A \code{2} x \code{D} x \code{F} array in column-major ordering
containing the spatial gradient in expression for each of
\code{D} latent variables at every triangle in the mesh.
Average of the vertices (3 for full triangles, 2 for degenerate triangles).}
\item{edges_pts}{A \code{2} x \code{D} x \code{E} array in column-major ordering
containing the spatial gradient in expression for each of
\code{D} latent variables at every primal edge (point-to-point) in the mesh.
Average of the two endpoints.}
\item{edges_tris}{A \code{2} x \code{D} x \code{E} array in column-major ordering
containing the spatial gradient in expression for each of
\code{D} latent variables at every dual edge (triangle-to-triangle) in the mesh.
Average of the two adjacent triangles.}
}
\description{
First, spatial gradients for each embedding dimension are computed for every
points (cell) by looking at each cell's neighbors. These point gradients can
be smoothed using bilateral/anisotropic filtering. Then gradients for triangles
are computed as the average of their vertices (3 for proper triangles, 2 for
degenerate exterior triangles at the boundaries). Finally, primal and dual
edge gradients are computed as the sum of the gradients for the two points and
two triangles, respectively, that are associated with each edge.
}
