% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_gradients_edges.R
\name{compute_gradients_edges}
\alias{compute_gradients_edges}
\title{Compute spatial gradient field for input to DMT}
\usage{
compute_gradients_edges(
  dmt,
  smooth_distance = "projected",
  smooth_similarity = "euclidean",
  smooth_iter = 1,
  edge_from_tri = FALSE,
  edge_from_pt = FALSE
)
}
\description{
Compute spatial gradient field for input to DMT
}
