% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_ui.R
\name{ConsolidateResults}
\alias{ConsolidateResults}
\title{Consolidate Tessera results from multiple samples (groups) after constructing
Tessera tiles separately on cells from each group.}
\usage{
ConsolidateResults(res, group.by)
}
\arguments{
\item{res}{Output of running GetTiles (when consolidate == FALSE).}

\item{group.by}{Name of metadata variable that identifies distinct groups.}
}
\description{
Consolidate Tessera results from multiple samples (groups) after constructing
Tessera tiles separately on cells from each group.
}
