% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tessera.R
\name{count.aneuploid}
\alias{count.aneuploid}
\title{count the number of aneuploid cells for the given chromosome}
\usage{
count.aneuploid(embryo, chromosome)
}
\arguments{
\item{embryo}{the embryo}

\item{chromosome}{the chromosome to test}
}
\value{
the number of aneuploid cells
}
\description{
count the number of aneuploid cells for the given chromosome
}
