% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cygnus-package.R
\docType{package}
\name{cygnus}
\alias{cygnus-package}
\alias{cygnus}
\title{cygnus}
\description{
Spatial tiling
}
\author{
\strong{Maintainer}: Ilya Korsunsky \email{ilya.korsunsky@gmail.com} (\href{https://orcid.org/0000-0003-4848-3948}{ORCID})

Authors:
\itemize{
  \item Daniel Stein
}

}
