% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{distanceCalculate}
\alias{distanceCalculate}
\title{Helper function for Moran's I.}
\usage{
distanceCalculate(x1, y1, x2, y2)
}
\arguments{
\item{x1}{First x-coordinate.}

\item{y1}{First y-coordinate.}

\item{x2}{Second x-coordinate.}

\item{y2}{Second y-coordinate.}
}
\value{
Euclidean distance between (x1, y1) and (x2, y2).
}
\description{
Helper function for Moran's I.
}
\author{
Matthew Cooper.
}
