% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_dmt.R
\name{dmt_assign_tiles}
\alias{dmt_assign_tiles}
\title{Assign points to tiles after DMT}
\usage{
dmt_assign_tiles(dmt)
}
\arguments{
\item{dmt}{A DMT data structure with \code{edges}, \code{pts}, and \code{e_sep} attributes.}
}
\value{
A DMT data structure with the following additional attributes:
\itemize{
\item \code{pts$agg_id}: Unique ID for the tile that each point belongs to.
\item \code{edges$agg_from}: Unique ID for the tile that \code{edges$from_pt} belongs to.
\item \code{edges$agg_to}: Unique ID for the tile that \code{edges$to_pt} belongs to.
}
}
\description{
Tiles are defined as points that are in the same connected component
in the mesh after discounting separatrix edges.
}
