% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_dmt.R
\name{dmt_get_separatrices}
\alias{dmt_get_separatrices}
\title{Get separatrices that separate points into components with strong boundaries}
\usage{
dmt_get_separatrices(dmt)
}
\arguments{
\item{dmt}{A DMT data structure with the following attributes:
\itemize{
\item \code{edges}: Data structure of edges in the mesh.
\item \code{prim}: Data structure of primal minimum spanning forest.
\item \code{dual}: Data structure of dual maximum spanning forest.
}}
}
\value{
A length \code{num_sep_edges} vector of edges (1-indexed) that make up the separatrices
that separate points into different components.
}
\description{
Finds the the collection of edges that lie along separatrices. These are found by
tracing paths from saddles to critical triangles (maxima) in the dual spanning forest.
Because the dual spanning forest is a maximum spanning forest on triangles in the mesh,
these paths will follow ridges, separating points into components with the strongest boundaries.
}
\details{
Saddle edges are found by interesting the saddle edges found in the primal forest with
saddles edges found in the dual forest. All boundary edges in the dual forest are also
included as possible saddle edges.
}
