% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_dmt.R
\name{dmt_set_f}
\alias{dmt_set_f}
\title{Set DMT scalar field values as the Frobenius norm of the total derivative}
\usage{
dmt_set_f(dmt, field)
}
\arguments{
\item{field}{A gradient field with the compressed representations \code{pts_svd},
\code{tris_svd}, \code{edges_pts_svd}, and \code{edges_tris_svd}. Each of these
is a \code{N} x \code{6} matrix with the following columns for each location:
\itemize{
\item \verb{dx_grad,dy_grad}: x,y directions of unit vector in the
direction of greatest change (first singular vector).
\item \verb{dx_ortho,dy_ortho}: x,y directions of unit vector orthogonal
to the direction of greatest change (second singular vector).
\item \verb{len_grad,len_ortho}: Magnitude of directional derivative in the
gradient and orthogonal directions (singular values).
}}
}
\value{
\code{dmt} with the following additional attributes:
\code{dmt$pts$f}, \code{dmt$tris$f}, \code{dmt$edges$f_prim}, and \code{dmt$edges$f_dual}
}
\description{
Set DMT scalar field values as the Frobenius norm of the total derivative
}
