% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_pca.R
\name{do_pca}
\alias{do_pca}
\title{Compute PCA embeddings from a raw counts matrix}
\usage{
do_pca(counts, npcs)
}
\arguments{
\item{counts}{A \code{n_genes} x \code{n_cells} counts matrix.
Must be convertible to dgCMatrix.}

\item{npcs}{Number of PCs to compute.}
}
\value{
A list with two features:
\itemize{
\item \code{loadings}: A \code{n_genes} x \code{npcs} matrix of gene loadings
for each PC. Each column is a unit vector.
\item \code{embeddings}: A \code{n_cells} x \code{npcs} matrix of cell
embeddings across all PCs. Each column \emph{j} has magnitude
equal to the \emph{j}th singular value. That is, PCs with
larger contribution to the total variance will have
embeddings of proportionally larger magnitude.
}
}
\description{
First, log-normalizes and Z-scores the counts matrix
and then performs PCA using SVD.
}
