% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tessera.R
\name{.bit.value}
\alias{.bit.value}
\title{Get the bit value for the given chromosome}
\usage{
.bit.value(chromosome)
}
\arguments{
\item{chromosome}{the chromosome to test}
}
\value{
the bit value
}
\description{
Get the bit value for the given chromosome
}
