% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tessera.R
\name{.count.empty.blocks}
\alias{.count.empty.blocks}
\title{Count the number of potential seed locations}
\usage{
.count.empty.blocks(d, chromosome)
}
\arguments{
\item{d}{the embryo}

\item{chromosome}{the chromosome to test}
}
\value{
the number of potential seeds
}
\description{
Cells that are not adjacent to a seed
}
