% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tessera.R
\name{.create.blank.sphere}
\alias{.create.blank.sphere}
\title{Create a sphere of evenly spaced cells}
\usage{
.create.blank.sphere(n.points)
}
\arguments{
\item{n.points}{the number of cells in the embryo}
}
\value{
a data frame with coordinates
}
\description{
Create a sphere of evenly spaced cells
}
