% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/E_step.R
\name{E_step_etahat}
\alias{E_step_etahat}
\title{Compute the Expectation of the random effects eta.
Part of the E-Step in the EM algorithm.
Follows Clayton and Kaldor (1987) for the Poisson-Gaussian model.}
\usage{
E_step_etahat(Vhat, Q, tau2, beta_hat, X, z, N)
}
\arguments{
\item{Vhat}{Estimated covariance matrix of eta.}

\item{Q}{Unscaled precision matrix.}

\item{tau2}{Precision/covariance matrix scaling.}

\item{beta_hat}{Current estimate of covariate effects.}

\item{X}{Covariate matrix.}

\item{z}{Observed counts.}

\item{N}{Library size or expected counts.}
}
\value{
Estimated mean of eta.
}
\description{
Compute the Expectation of the random effects eta.
Part of the E-Step in the EM algorithm.
Follows Clayton and Kaldor (1987) for the Poisson-Gaussian model.
}
\note{
Applies to a single area.

This method is an Empirical Bayes approximation.
}
\author{
Florica J Constantine, florica AT berkeley.edu
}
