% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/E_step.R
\name{E_step_predict}
\alias{E_step_predict}
\title{Compute the best fit to the data.
Associated with the E-Step in the EM algorithm, but not needed to run.
I.e., this is a utility function.
Follows Clayton and Kaldor (1987) for the Poisson-Gaussian model.}
\usage{
E_step_predict(theta_hat, N)
}
\arguments{
\item{theta_hat}{Estimated poisson parameters.}

\item{N}{Library size or expected counts.}
}
\value{
Estimated values z (model fits).
}
\description{
Compute the best fit to the data.
Associated with the E-Step in the EM algorithm, but not needed to run.
I.e., this is a utility function.
Follows Clayton and Kaldor (1987) for the Poisson-Gaussian model.
}
\author{
Florica J Constantine, florica AT berkeley.edu
}
