% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/E_step.R
\name{E_step_thetahat_PLU}
\alias{E_step_thetahat_PLU}
\title{Compute the estimated poisson parameters theta.
Note that eta = log theta.
Associated with the E-Step in the EM algorithm, but not needed to run.
I.e., this is a utility function.
Follows Clayton and Kaldor (1987) for the Poisson-Gaussian model.}
\usage{
E_step_thetahat_PLU(decomp, eta_hat)
}
\arguments{
\item{decomp}{Estimated covariance matrix of eta PLU decomposition.}

\item{eta_hat}{Estimated mean of eta.}
}
\value{
Estimated poisson parameters theta_hat.
}
\description{
Compute the estimated poisson parameters theta.
Note that eta = log theta.
Associated with the E-Step in the EM algorithm, but not needed to run.
I.e., this is a utility function.
Follows Clayton and Kaldor (1987) for the Poisson-Gaussian model.
}
\note{
Applies to a single area.
}
\author{
Florica J Constantine, florica AT berkeley.edu
}
