% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{estimate_field_edges_cpp}
\alias{estimate_field_edges_cpp}
\title{Compute a spatial gradient field along each edge}
\usage{
estimate_field_edges_cpp(coords, embeddings, from_pt, to_pt)
}
\arguments{
\item{coords}{A \code{N} x \code{2} matrix of cell coordinates.}

\item{embeddings}{A \code{N} x \code{D} matrix of cell embeddings.}

\item{from_pt, to_pt}{A pair of \code{E}-length vectors indicating
the start and end points of each edge (0-indexed).}
}
\value{
A \code{2} x \code{D} x \code{E} array in column-major ordering
containing the spatial gradient in expression for each of
\code{D} embedding dimensions at every edge.
}
\description{
Distance between neighboring cells is normalized to unit
distance so that only the orientation of each edge matters.
The gradient is then the difference in expression of each
embedding dimension between the two endpoints of the edge.
}
