% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\name{example_prepData.rds}
\alias{example_prepData.rds}
\title{Pre-processed TESSERA Data Object}
\format{
A \code{.rds} file containing the processed TESSERA data structure,
including filtered counts and spatial weights formatted for the Leroux model.
}
\description{
An object of class \code{TESSERAData} (or the list format expected by
the package) generated by passing raw data through \code{prep_data()}.
}
