% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\name{example_TESSERA_out_Leroux.rds}
\alias{example_TESSERA_out_Leroux.rds}
\title{Fitted TESSERA Leroux Model Object}
\format{
A \code{.rds} file containing:
\describe{
\item{beta_hat}{Estimated fixed effects.}
\item{tau2_hat}{Estimated spatial variance.}
\item{gamma_hat}{Estimated spatial correlation (Leroux parameter).}
\item{...}{Other tracked quantities.}
}
}
\description{
A fitted model object resulting from running \code{TESSERA_lattice}
on the example prepared data.
}
