% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{findDuplicates}
\alias{findDuplicates}
\title{Find duplicates within a vector}
\usage{
findDuplicates(input)
}
\arguments{
\item{input}{Vector of values.}
}
\value{
Vector of values from \code{input} that appear at least twice.
Each value that appears N times in \code{input} will appear N-1 times in the output.
}
\description{
Find duplicates within a vector
}
