% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_analysis.R
\name{FindSubClusterCustom}
\alias{FindSubClusterCustom}
\title{Find sub-clusters within a specific cluster of a Seurat object}
\usage{
FindSubClusterCustom(
  obj,
  cluster,
  clusters.name = "seurat_clusters",
  sub.clusters.name = NULL,
  resolution = 0.5,
  algorithm = 1,
  npcs = 30,
  method = "igraph",
  n_neighbors = 15,
  fast_sgd = TRUE,
  scale.factor = NULL,
  use.existing.embeddings = NULL,
  meta.vars.include = NULL,
  harmony.group.by.vars = NULL,
  early_stop = TRUE,
  return_obj_sub = FALSE,
  ...
)
}
\arguments{
\item{obj}{A Seurat object.}

\item{cluster}{The cluster label to sub-cluster.}

\item{clusters.name}{Name of the metadata column in \code{obj} containing the cluster labels.}

\item{sub.clusters.name}{Name of the metadata column to store the sub-cluster labels.
If NULL, defaults to \verb{<clusters.name>_s<cluster>}.}

\item{resolution}{Resolution parameter for clustering.}

\item{algorithm}{Clustering algorithm to use. See \code{?Seurat::FindClusters} for details.}

\item{npcs}{Number of principal components to use for sub-clustering.}

\item{method}{Method to use for clustering. See \code{?Seurat::FindClusters} for details.}

\item{n_neighbors}{Number of neighbors to use in UMAP.}

\item{fast_sgd}{Whether to use fast SGD in UMAP.}

\item{scale.factor}{Scale factor for normalization. If NULL, use median nCount_RNA.}

\item{use.existing.embeddings}{Name of existing dimensional reduction in \code{obj}
to use for sub-clustering. If NULL, compute PCA on the subsetted data.}

\item{meta.vars.include}{Metadata variables to include in the sub-clustered object.}

\item{harmony.group.by.vars}{Metadata variables to use for Harmony integration.}

\item{early_stop}{Whether to use early stopping in Harmony.}

\item{return_obj_sub}{If TRUE, return a list with the updated \code{obj} and the sub-clustered object.}

\item{...}{Additional parameters to pass to \code{harmony::RunHarmony}. Ignored if \code{harmony.group.by.vars} is NULL.}
}
\value{
The input Seurat object with sub-cluster labels added to metadata.
}
\description{
Find sub-clusters within a specific cluster of a Seurat object
}
