% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_analysis.R
\name{FindTileSubCluster}
\alias{FindTileSubCluster}
\title{Find sub-clusters within a specific cluster of a Seurat object using matching tile and cell data}
\usage{
FindTileSubCluster(
  obj,
  cluster,
  cell_obj = NULL,
  clusters.name = "seurat_clusters",
  tile_assay = NULL,
  cell_assay = NULL,
  sub.clusters.name = NULL,
  resolution = 0.2,
  algorithm = 4,
  npcs = 20,
  method = "igraph",
  n_neighbors = 25,
  fast_sgd = TRUE,
  scale.factor = NULL,
  use.existing.embeddings = NULL,
  smooth_emb = c(0, 1),
  graph.name.cells = "cell_adj",
  meta.vars.include = NULL,
  harmony.group.by.vars = NULL,
  early_stop = TRUE,
  return_obj_sub = FALSE
)
}
\arguments{
\item{obj}{A Seurat object containing tile-level data.}

\item{cluster}{The cluster label to sub-cluster.}

\item{cell_obj}{(Opional) A Seurat object containing cell-level data with a \code{tile_id} metadata column.
If NULL, sub-clustering is performed using only \code{obj} without recomputing tile embeddings from cell data.}

\item{clusters.name}{Name of the metadata column in \code{obj} containing the cluster labels.}

\item{tile_assay}{Assay in \code{obj} to use for sub-clustering. If NULL, use DefaultAssay.}

\item{cell_assay}{Assay in \code{cell_obj} to use for computing cell embeddings. If NULL, use DefaultAssay.}

\item{sub.clusters.name}{Name of the metadata column to store the sub-cluster labels.
If NULL, defaults to \verb{<clusters.name>_s<cluster>}.}

\item{resolution}{Resolution parameter for clustering.}

\item{algorithm}{Clustering algorithm to use. See \code{?Seurat::FindClusters} for details.}

\item{npcs}{Number of principal components to use for sub-clustering.}

\item{method}{Method to use for clustering. See \code{?Seurat::FindClusters} for details.}

\item{n_neighbors}{Number of neighbors to use in UMAP.}

\item{fast_sgd}{Whether to use fast SGD in UMAP.}

\item{scale.factor}{Scale factor for normalization. If NULL, use median nCount_RNA.}

\item{use.existing.embeddings}{Name of existing dimensional reduction in \code{obj}
to use for sub-clustering. If NULL, compute PCA on the subsetted cell-level data.}

\item{smooth_emb}{Number of smoothing iterations to perform on the cell embeddings,
as for \code{GetTiles}. If a vector, then embeddings after each specified
iteration are concatenated. If \code{0} is included, then the original embeddings are also included.}

\item{graph.name.cells}{Name of the graph in \code{cell_obj} containing the cell adjacency graph.}

\item{meta.vars.include}{Metadata variables to include in the sub-clustered object.}

\item{harmony.group.by.vars}{Metadata variables to use for Harmony integration.}

\item{early_stop}{Whether to use early stopping in Harmony.}

\item{return_obj_sub}{If TRUE, return a list with the updated \code{obj} and the sub-clustered object.}
}
\value{
The input Seurat object with sub-cluster labels added to metadata.
}
\description{
Find sub-clusters within a specific cluster of a Seurat object using matching tile and cell data
}
