% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/post_analysis.R
\name{fit_empirical_null_fdrtool}
\alias{fit_empirical_null_fdrtool}
\title{Wrapper for Empirical null estimation and p-value computation}
\usage{
fit_empirical_null_fdrtool(wald_stats)
}
\arguments{
\item{wald_stats}{A numeric vector of Wald statistics.
Note: This function expects signed t-statistics, or, the un-squared statistics.}
}
\value{
A list containing the following components:
\itemize{
\item \strong{wald_pval}: The estimated (unadjusted) p-values.
\item \strong{params}: The estimated parameters of the empirical null distribution;
the $params output from \code{fdrtool::fdrtool}.
\item \strong{fndr_out}: The output object from \code{fdrtool::fdrtool}.
}
}
\description{
In generalized linear mixed models (GLMMs) and the TESSERA framework,
theoretical null distributions for Wald statistics are often unreliable
due to finite-sample biases and E-step approximations. This function
is a wrapper for the fdrtool fndr method, which assumes that the null is a
centered normal distribution with some variance (estimated).
}
\examples{
# Simulate null and non-null Wald statistics
null_stats <- rnorm(n = 2000, mean = 0.1, sd = 4)
alt_stats <- rnorm(n = 500, mean = 5, sd = 2)
wald_vec <- c(null_stats, alt_stats)

# Select optimal threshold
fit_empirical_null_fdrtool(wald_vec)
}
\author{
Florica J Constantine, florica AT berkeley.edu
}
