% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/post_analysis.R
\name{fit_scaled_noncentral_chi2_DEoptim}
\alias{fit_scaled_noncentral_chi2_DEoptim}
\title{Fit a scaled non-central chi^2_1 distribution to a vector of statistics.
The fit is conditional, below a threshold.}
\usage{
fit_scaled_noncentral_chi2_DEoptim(wald_stats, wald_thresh)
}
\arguments{
\item{wald_stats}{Vector of Wald statistics: non-negative F-statistics.
Square t-statistics to obtain F-statistics.}

\item{wald_thresh}{Threshold below which we fit.}
}
\value{
Vector of (scaling, shift).
(scaling) chi_1^2(shift).
}
\description{
We numerically maximize the likelihood.
}
\note{
Consider using the trimfrac parameter if the data have outliers as
this function is sensitive to the presence of large values/outliers.
}
\examples{
fit_scaled_noncentral_chi2_DEoptim(5 * stats::rchisq(1000, 1, ncp=10), 20)
}
\author{
Florica J Constantine, florica AT berkeley.edu
}
