% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/post_analysis.R
\name{fit_scaled_noncentral_chi2_mom}
\alias{fit_scaled_noncentral_chi2_mom}
\title{Fit a scaled non-central chi^2_1 distribution to a vector of statistics.
The fit is conditional, below a threshold.}
\usage{
fit_scaled_noncentral_chi2_mom(wald_stats, wald_thresh)
}
\arguments{
\item{wald_stats}{Vector of Wald statistics: non-negative F-statistics.}

\item{wald_thresh}{Threshold below which we fit.}
}
\value{
Vector of (scaling, shift).
}
\description{
Fit a scaled non-central chi^2_1 distribution to a vector of statistics.
The fit is conditional, below a threshold.
}
\examples{
fit_scaled_noncentral_chi2_mom(5 * stats::rchisq(10000, 1, ncp=10), 200)
}
\author{
Florica J Constantine, florica AT berkeley.edu
}
