% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/post_analysis.R
\name{fit_scaled_noncentral_chi2_optimx}
\alias{fit_scaled_noncentral_chi2_optimx}
\title{Fit a scaled non-central chi^2_1 distribution using BOBYQA}
\usage{
fit_scaled_noncentral_chi2_optimx(wald_stats, wald_thresh)
}
\arguments{
\item{wald_stats}{Vector of Wald statistics (non-negative).}

\item{wald_thresh}{Threshold below which data is included in the fit.}
}
\description{
Numerically maximizes the likelihood of a scaled non-central chi-square
distribution (df=1) truncated above a specific threshold.
}
\examples{
fit_scaled_noncentral_chi2_optimx(5 * stats::rchisq(1000, 1, ncp=10), 20)
}
