% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{get_e_sep}
\alias{get_e_sep}
\title{Get the collection of edges that lie along separatrices}
\usage{
get_e_sep(epaths, saddles, nedges)
}
\arguments{
\item{epaths}{A list of length \code{2*num_saddles} containing the two paths from each saddle edge to the
two critical points that it joins. Each path is a numeric vector of edge indices (1-indexed).}

\item{saddles}{A length \code{num_saddles} vector with edge indices for saddle edges (1-indexed).}

\item{nedges}{Total number of edges in the mesh.}
}
\value{
A length \code{num_sep_edges} vector of edges (0-indexed) that are saddle edges in \code{saddles} or
lie along the paths in \code{epaths}. These edges make up the separatrices that separate points into
different components.
}
\description{
Get the collection of edges that lie along separatrices
}
