% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_ui.R
\name{GetTiles}
\alias{GetTiles}
\title{Generic function that runs the Tessera algorithm on single-cell spatial data}
\usage{
GetTiles(...)
}
\arguments{
\item{...}{
  Arguments passed on to \code{\link[=GetTiles.default]{GetTiles.default}}
  \describe{
    \item{\code{meta_vars_include}}{Names of columns in meta_data to include in \code{dmt$pts}.}
    \item{\code{group.by}}{Name of column in \code{meta_data} that provides the group IDs. Tessera tiles are
constructed separately for each group (which could be separate experimental samples or FOVs).}
    \item{\code{npcs}}{Number of PCs to compute for input to segmentation.}
    \item{\code{prune_thresh_quantile}}{Floating point value between 0 and 1, inclusive.
Quantile of edge length above which edges are pruned. Defaults to 0.95.}
    \item{\code{prune_min_cells}}{Minimum number of cells required for a connected
component of triangles to be kept. Defaults to 10.}
    \item{\code{prune_thresh}}{Edge length above which edges are pruned. If equal to NA,
then this value is ignored and \code{thresh_quantile} is used to compute
the threshold. Otherwise, if \code{thresh} is set, then \code{thresh_quantile}
is ignored. Defaults to NA.}
    \item{\code{smooth_distance}}{One of \code{c('none', 'euclidean', 'projected', 'constant')}.
If either \code{smooth_distance} or \code{smooth_similarity} is \code{'none'},
then no smoothing of the gradient field is conducted. Defaults to \code{'projected'}.}
    \item{\code{smooth_similarity}}{One of \code{c('none', 'euclidean', 'projected', 'constant')}.
If either \code{smooth_distance} or \code{smooth_similarity} is \code{'none'},
then no smoothing of the gradient field is conducted. Defaults to \code{'projected'}.}
    \item{\code{smooth_iter}}{Number of rounds of gradient smoothing.}
    \item{\code{max_npts}}{Maximum number of cells allowed in each tile during the
agglomerative clustering phase.}
    \item{\code{min_npts}}{Minimum number of cells allowed in each tile during the
agglomerative clustering phase.}
    \item{\code{alpha}}{Parameter for scoring transcriptional similarity between adjacent tiles during
the agglomerative clustering phase. For \code{alpha}, 0.2 = conservative merging, 2 = liberal merging.}
    \item{\code{future.globals.maxSize}}{Maximum allowed size (in bytes) of global variables that are exported to each parallel worker.
Increase this value if you get an error about global object size. Default is 8*1024^3 (8 GB).}
    \item{\code{consolidate}}{Whether to consolidate results from multiple groups into a single collection of
points and tiles (TRUE) or to return a list of separate results for each group (FALSE).}
    \item{\code{verbose}}{Whether to print progress messages for each stage of the segmentation pipeline.}
  }}
}
\value{
If used with a Seurat object, it will return a pair of Seurat objects:
\enumerate{
\item the input single-cell object updated with tile assignments for each cell, and
\item a Seurat object where each item represents an individual Tessera tile.
}

For standalone operation, it returns Lists with the output of Tessera segmentation
(see \code{\link[=GetTiles.default]{GetTiles.default()}}).
}
\description{
GetTiles is a generic function that runs the main Tessera algorithm.
If working with a Seurat object, please refer to the documentation of
the appropriate generic API: \code{\link[=GetTiles.Seurat]{GetTiles.Seurat()}}. If users work with other
forms of the input, they can pass them directly to Tessera using the
\code{\link[=GetTiles.default]{GetTiles.default()}} API. The function arguments listed here are common in all
GetTiles interfaces.
}
\seealso{
Other GetTiles: 
\code{\link{GetTiles.Seurat}()},
\code{\link{GetTiles.default}()}
}
\concept{GetTiles}
