% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/comparisons.R
\name{glm_nb_wrapper}
\alias{glm_nb_wrapper}
\title{Wrapper to fit basic Negative Binomial GLM.}
\usage{
glm_nb_wrapper(z_list, X_list)
}
\arguments{
\item{z_list}{List of vectors of observed counts---one vector per area.}

\item{X_list}{List of design or covariate matrices---one matrix per area.
Same length and ordering as z_list.
Matrices with number of rows equal to length of corresponding vector in
z_list.}
}
\value{
The output list from glm.nb, plus a time field for how long the
function ran for.
}
\description{
This function theoretically works for multiple areas by stacking everything
together and ignoring spatial correlation.
}
\note{
Depends on MASS.
}
\author{
Florica J Constantine, florica AT berkeley.edu
}
