% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_initdata.R
\name{init_data}
\alias{init_data}
\title{Generate mesh data structure from coordinates, for input to DMT analysis}
\usage{
init_data(X, Y, counts, meta_data = NULL, meta_vars_include = c())
}
\arguments{
\item{X, Y}{A pair of numeric vectors with the coordinates for each of V points.}

\item{counts}{A G x V gene-by-cell matrix of transcript counts.}

\item{meta_data}{A data frame with additional cell metadata to include.}

\item{meta_vars_include}{Names of columns in meta_data to include.}
}
\value{
A list containing the mesh data structures:
\itemize{
\item \code{pts} is a V-by-M data table with columns \code{X} and \code{Y} containing the
coordinates of cells and additional metadata.
\item \code{tris} is a F-by-4 data table containing the X,Y coordinates of each
triangle's centroid in the first two columns, and area and
largest height of each triangle in the last two columns.
\item \code{edges} is a E-by-14 data table with columns \code{from_pt}, \code{to_pt}, \code{from_tri}, \code{to_tri},
\code{x0_pt}, \code{x1_pt}, \code{y0_pt}, \code{y1_pt}, \code{x0_tri}, \code{x1_tri}, \code{y0_tri}, \code{y1_tri},
\code{length_pt}, \code{length_tri}. If only one triangle uses an edge, then the \code{from_tri},
\code{x0_tri}, and \code{y0_tri} fields will contain NaN values.
\item \code{tri_to_pt} is a F-by-V sparse matrix with value 1 at (i,j) if
triangle i uses point j as a vertex.
\item \code{counts} is a G x V gene-by-cell matrix of transcript counts.
}

Note that all indices stored in these data structures are 1-indexed.
}
\description{
Creates a mesh from point coordinates using Delauney triangulation.
Stores the points, triangles, and edges for the mesh, as well
as a mapping from each triangle to its associated vertices.
A gene-by-cell matrix can also be included, as well as additional
metadata for each point.
}
