% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{init_edges_cpp}
\alias{init_edges_cpp}
\title{Calculates triangles' centroids, areas, and heights from vertices}
\usage{
init_edges_cpp(triplets, pts, tris)
}
\arguments{
\item{triplets}{A M-by-3 matrix with indices for the points that
correspond each triangle's vertices, where M is the number of triangles.}

\item{pts}{A N-by-2 matrix with indices for the X,Y coordinates of
each point.}

\item{tris}{A M-by-4 matrix containing the X,Y coordinates of each
triangle's centroid in the first two columns, and area and
largest height of each triangle in the last two columns.}
}
\value{
A E-by-14 matrix with columns \code{from_pt}, \code{to_pt}, \code{from_tri}, \code{to_tri},
\code{x0_pt}, \code{x1_pt}, \code{y0_pt}, \code{y1_pt}, \code{x0_tri}, \code{x1_tri}, \code{y0_tri}, \code{y1_tri},
\code{length_pt}, \code{length_tri}. If only one triangle uses an edge, then the \code{from_tri},
\code{x0_tri}, and \code{y0_tri} fields will contain NaN values.

Note that indices that reference the \code{pts} and \code{tris} tables are 1-indexed.
}
\description{
Calculates triangles' centroids, areas, and heights from vertices
}
