% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_aggregation.R
\name{init_scores}
\alias{init_scores}
\title{Initialize tile scores for aggregation}
\usage{
init_scores(aggs, agg_mode, ...)
}
\arguments{
\item{aggs}{A tile data structure.}

\item{agg_mode}{Method to use to calculate aggregation scores (1, 2, or 3).}

\item{...}{Additional parameters for different modes:
\enumerate{
\item No additional parameters.
\item Requires \code{alpha} and \code{max_npts}. For \code{alpha}, 0.2 = conservative merging, 2 = liberal merging.
\item Requires \code{alpha} and \code{min_npts}. For \code{alpha}, 0.2 = conservative merging, 2 = liberal merging.
}}
}
\value{
\code{aggs} with scores for aggregation stored in attributes:
\item{edges}{Additional attributes are calculated:
* \code{dscore}: Overall score for merging two tiles. Product of \code{w}, \code{score_size}, and \code{dC}.
* \code{w}: Gene expression similarity score.
* \code{score_size}: Penalizes tiles with many points.
* \code{perimeter_merge}: Perimeter of merged tile.
}
\item{d_mu,d_sig}{Parameters used to calculate \code{w}.}
\item{pcs_merged}{Average PCs for merged tile.}
}
\description{
Higher scores favor merging.
}
\details{
Methods for different modes of aggregation:
\enumerate{
\item \code{dscore} has already been manually calculated and stored in \code{aggs$edges$dscore}.
Will set the \code{score}, \code{score_size}, and \code{compactness} attributes of \code{aggs$meta_data} to 0.
\item \code{dscore} is the product of three factors:
\itemize{
\item \code{w}: A 2-cluster GMM is used to determine the mean \code{mu} and standard deviation \code{sig} of the distance
between tiles that have similar gene expression (Euclidean distance \code{d} in PC space).
Then we define \code{d_mu = mu + sig} and \code{d_sig = alpha * sig}, and calculate \code{w} as
\code{w = 0.5 - 1 / (1 + exp(-(d - d_mu) / d_sig))}. Ranges from -0.5 to 0.5. If adjacent tiles
are very dissimilar (\verb{d >> d_mu}), then \code{d} is large, and \code{w} is close to \code{-0.5}. If adjacent
tiles are very similar (\code{d < d_mu}), then \code{d} is small, and \code{w} is positive.
\item \code{score_size}: \code{(1 - npts_from/max_npts) * (1 - npts_to/max_npts)}. Ranges from 0 to 1.
If merging the two tiles would have a total number of points ≥\code{max_npts}, then \code{score_size} is \code{-Inf},
which prevents merging.
\item \code{dC}: \code{.5 * (C_merge - C_from - C_to + 1)}. Ranges from 0 to 1.
}
\item \code{dscore} is the product of same three factors as mode 2, but \code{score_size} has already been precomputed.
Additionally, \code{dscore} is set to -1 if both adjacent tiles have at least \code{min_npts} cells.
}
}
