% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/comparisons.R
\name{INLA_BYM2_wrapper}
\alias{INLA_BYM2_wrapper}
\title{Wrapper to fit a multi-sample BYM2 model with INLA (INLA).}
\usage{
INLA_BYM2_wrapper(
  z_list,
  X_list,
  W_list,
  model_family = "poisson",
  z_offset = 0.5,
  num_threads = 1,
  compute_dic = TRUE,
  compute_waic = TRUE,
  compute_cpo = FALSE,
  verbose = TRUE,
  library_size_list = NULL
)
}
\arguments{
\item{z_list}{List of vectors of observed counts---one vector per area.}

\item{X_list}{List of design or covariate matrices---one matrix per area.
Same length and ordering as z_list.
Matrices with number of rows equal to length of corresponding vector in
z_list.}

\item{W_list}{List of spatial adjacency matrices---one matrix per area.}

\item{model_family}{"poisson" or "gaussian".
Which model family to fit. Note that Poisson models don't scale well in the
multi-area setting---the spline is fit on a per-area basis.}

\item{z_offset}{In the "gaussian" case, the counts z are transformed as
log(z + z_offset).}

\item{num_threads}{Number threads to run.}

\item{compute_dic}{INLA parameter.}

\item{compute_waic}{INLA parameter.}

\item{compute_cpo}{INLA parameter.}

\item{verbose}{Whether INLA is verbose.}

\item{library_size_list}{A list with a vector of library sizes for each sample.
Total counts for each location if there are more than one gene/measurement;
if NULL, set to 1.}
}
\value{
The output list from INLA::inla, plus a time field
for how long the function ran for.
}
\description{
This function works for multiple areas, but is very slow.
}
\note{
Requires the INLA library.
}
\author{
Florica J Constantine, florica AT berkeley.edu
}
