% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{inv_svd_field_cpp}
\alias{inv_svd_field_cpp}
\title{Inverse SVD transform to reconstruct spatial gradient field}
\usage{
inv_svd_field_cpp(svd_list)
}
\arguments{
\item{svd_list}{A list with elements \code{u} (\code{2} x \code{2} x \code{N}), \code{s} (\code{2} x \code{N}), and
\code{v} (\code{D} x \code{2} x \code{N}) containing the left singular vectors, singular values,
and right singular vectors for each point (or edge/triangle).}
}
\value{
A \code{2} x \code{D} x \code{N} array in column-major ordering
containing the spatial gradient in expression for each of
\code{D} embedding dimensions at every point (or edge/triangle).
}
\description{
Inverse SVD transform to reconstruct spatial gradient field
}
