% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/post_analysis.R
\name{inversePrecisionMatrixWald}
\alias{inversePrecisionMatrixWald}
\title{Helper function for Wald stat covariances.}
\usage{
inversePrecisionMatrixWald(A)
}
\arguments{
\item{A}{Estimated precision matrix.}
}
\value{
Inverse of A.
}
\description{
Invert a precision matrix, handling failure cases.
}
\author{
Florica J Constantine, florica AT berkeley.edu
}
