% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tessera.R
\name{is.aneuploid}
\alias{is.aneuploid}
\title{Test if the given chromosome in the given cell is aneuploid}
\usage{
is.aneuploid(embryo, cell.index, chromosome)
}
\arguments{
\item{embryo}{the embryo}

\item{cell.index}{the cell to test (0 for all cells)}

\item{chromosome}{the chromosome to test}
}
\value{
if cell.index is >0, return true if the chromosome is aneuploid,
false otherwise. If cell.index is 0, return a boolean vector of all cells.
}
\description{
Test if the given chromosome in the given cell is aneuploid
}
