% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/models.R
\name{kernel.sph}
\alias{kernel.sph}
\title{Spherical Kernel Function.}
\usage{
kernel.sph(d, sigma2, rho)
}
\arguments{
\item{d}{Vector of distances.}

\item{sigma2}{Scale parameter.}

\item{rho}{Range parameter.}
}
\value{
A vector of covariance values.
}
\description{
Spherical Kernel Function.
}
\note{
Applies to a single area.
}
\author{
Florica J Constantine, florica AT berkeley.edu
}
