% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/models.R
\name{kernel_matern}
\alias{kernel_matern}
\title{Matern Kernel Function.}
\usage{
kernel_matern(d, sigma2, rho, nu)
}
\arguments{
\item{d}{Vector of distances.}

\item{sigma2}{Scale parameter.}

\item{rho}{Range parameter.}

\item{nu}{(kappa) Order of kernel.}
}
\value{
A vector of covariance values.
}
\description{
Matern Kernel Function.
}
\note{
Applies to a single area.
}
\author{
Florica J Constantine, florica AT berkeley.edu
}
