% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/comparisons.R
\name{lm_wrapper}
\alias{lm_wrapper}
\title{Wrapper to fit basic linear model.}
\usage{
lm_wrapper(z_list, X_list, transform_z = TRUE, z_offset = 0.5)
}
\arguments{
\item{z_list}{List of vectors of observed counts---one vector per area.}

\item{X_list}{List of design or covariate matrices---one matrix per area.
Same length and ordering as z_list.
Matrices with number of rows equal to length of corresponding vector in
z_list.}

\item{transform_z}{Boolean: log-transform z or not.}

\item{z_offset}{If transform_z, the counts z are transformed as
log(z + z_offset).}
}
\value{
The output list from lm, plus a time field for how long the
function ran for.
}
\description{
This function theoretically works for multiple areas by stacking everything
together and ignoring spatial correlation.
}
\author{
Florica J Constantine, florica AT berkeley.edu
}
