% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/M_step.R
\name{M_step_BRISC}
\alias{M_step_BRISC}
\title{Holding beta constant, use BRISC to find the spatial parameters of eta - X beta.
Part of the M-Step in the EM algorithm.
ONLY APPLIES TO THE GAUSSIAN PROCESS MODEL.}
\usage{
M_step_BRISC(eta_hat, beta_hat, X, coords, cov_type = "Exp", k = 15)
}
\arguments{
\item{eta_hat}{Estimated mean of eta.}

\item{beta_hat}{Current estimate of covariate effects.}

\item{X}{Covariate matrix.}

\item{coords}{Matrix of (x, y) coordinates.}

\item{cov_type}{String for covariance model type.
"Exp", "Sph", "Gau", and "Mat" are supported.}

\item{k}{Number of neighbors.}
}
\value{
A vector of fitted parameters.
Nugget variance (partial sill), Spatial variance (partial sill),
Spatial range (scales distance); if Matern, smoothness kappa.
}
\description{
Holding beta constant, use BRISC to find the spatial parameters of eta - X beta.
Part of the M-Step in the EM algorithm.
ONLY APPLIES TO THE GAUSSIAN PROCESS MODEL.
}
\note{
Applies to a single area.

Requires the BRISC library (called indirectly).
}
\author{
Florica J Constantine, florica AT berkeley.edu
}
