% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/M_step.R
\name{M_step_tau2}
\alias{M_step_tau2}
\title{Maximize the expected likelihood in tau^2, holding other variables constant.
Part of the M-Step in the EM algorithm.}
\usage{
M_step_tau2(Vhat, eta_hat, Q, beta_hat, X)
}
\arguments{
\item{Vhat}{Estimated covariance matrix of eta.}

\item{eta_hat}{Estimated mean of eta.}

\item{Q}{Unscaled precision matrix.}

\item{beta_hat}{Current estimate of covariate effects.}

\item{X}{Covariate matrix.}
}
\value{
Maximum likelihood estimate of scaling parameter tau^2.
}
\description{
Maximize the expected likelihood in tau^2, holding other variables constant.
Part of the M-Step in the EM algorithm.
}
\note{
Applies to a single area.
}
\author{
Florica J Constantine, florica AT berkeley.edu
}
