% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/M_step.R
\name{M_step_variogram}
\alias{M_step_variogram}
\title{Holding beta constant, fit a variogram model to eta - X beta.
Part of the M-Step in the EM algorithm.
Instead of a traditional MLE, which has cubic time complexity, we
fit a variogram to estimate the kernel parameters, which is quadratic time.
ONLY APPLIES TO THE GAUSSIAN PROCESS MODEL.}
\usage{
M_step_variogram(eta_hat, beta_hat, X, coords, cov_type = "Exp")
}
\arguments{
\item{eta_hat}{Estimated mean of eta.}

\item{beta_hat}{Current estimate of covariate effects.}

\item{X}{Covariate matrix.}

\item{coords}{Matrix of (x, y) coordinates.}

\item{cov_type}{String for covariance model type.
"Exp", "Sph", "Gau", and "Mat" are supported.}
}
\value{
A vector of fitted parameters.
Nugget variance (partial sill), Spatial variance (partial sill),
Spatial range (scales distance); if Matern, smoothness kappa.
}
\description{
Holding beta constant, fit a variogram model to eta - X beta.
Part of the M-Step in the EM algorithm.
Instead of a traditional MLE, which has cubic time complexity, we
fit a variogram to estimate the kernel parameters, which is quadratic time.
ONLY APPLIES TO THE GAUSSIAN PROCESS MODEL.
}
\note{
Applies to a single area.

Requires the gstat library.
}
\author{
Florica J Constantine, florica AT berkeley.edu
}
