% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_analysis.R
\name{MakeSubClusterObj}
\alias{MakeSubClusterObj}
\title{Create a Seurat object for sub-clustering a specific cluster}
\usage{
MakeSubClusterObj(
  obj,
  cluster,
  clusters.name = "seurat_clusters",
  npcs = 30,
  fast_sgd = TRUE,
  n_neighbors = 15,
  assay = NULL,
  scale.factor = NULL,
  use.existing.embeddings = NULL,
  meta.vars.include = NULL,
  harmony.group.by.vars = NULL,
  early_stop = TRUE,
  ...
)
}
\arguments{
\item{obj}{A Seurat object.}

\item{cluster}{The cluster label to sub-cluster.}

\item{clusters.name}{Name of the metadata column in \code{obj} containing the cluster labels.}

\item{npcs}{Number of principal components to use for sub-clustering.}

\item{fast_sgd}{Whether to use fast SGD in UMAP.}

\item{n_neighbors}{Number of neighbors to use in UMAP.}

\item{scale.factor}{Scale factor for normalization. If NULL, use median nCount_RNA.}

\item{use.existing.embeddings}{Name of existing dimensional reduction in \code{obj}
to use for sub-clustering. If NULL, compute PCA on the subsetted data.}

\item{meta.vars.include}{Metadata variables to include in the sub-clustered object.}

\item{harmony.group.by.vars}{Metadata variables to use for Harmony integration.}

\item{early_stop}{Whether to use early stopping in Harmony.}

\item{...}{Additional parameters to pass to \code{harmony::RunHarmony}. Ignored if \code{harmony.group.by.vars} is NULL.}
}
\value{
A Seurat object for the sub-clustered cells.
}
\description{
Create a Seurat object for sub-clustering a specific cluster
}
