% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_analysis.R
\name{MakeTileSubClusterObj}
\alias{MakeTileSubClusterObj}
\title{Create a Seurat object for sub-clustering a specific cluster from matching tile and cell data}
\usage{
MakeTileSubClusterObj(
  tile_obj,
  cluster,
  cell_obj,
  clusters.name = "seurat_clusters",
  npcs = 30,
  tile_assay = NULL,
  cell_assay = NULL,
  fast_sgd = TRUE,
  n_neighbors = 15,
  scale.factor = NULL,
  use.existing.embeddings = NULL,
  smooth_emb = c(0, 1),
  graph.name.cells = "cell_adj",
  meta.vars.include = NULL,
  harmony.group.by.vars = NULL,
  early_stop = TRUE
)
}
\arguments{
\item{tile_obj}{A Seurat object containing tile-level data.}

\item{cluster}{The cluster label to sub-cluster.}

\item{cell_obj}{A Seurat object containing cell-level data with a \code{tile_id} metadata column.}

\item{clusters.name}{Name of the metadata column in \code{tile_obj} containing the cluster labels.}

\item{npcs}{Number of principal components to use for sub-clustering.}

\item{tile_assay}{Assay in \code{tile_obj} to use for sub-clustering. If NULL, use DefaultAssay.}

\item{cell_assay}{Assay in \code{cell_obj} to use for computing cell embeddings. If NULL, use DefaultAssay.}

\item{fast_sgd}{Whether to use fast SGD in UMAP.}

\item{n_neighbors}{Number of neighbors to use in UMAP.}

\item{scale.factor}{Scale factor for normalization. If NULL, use median nCount_RNA.}

\item{use.existing.embeddings}{Name of existing dimensional reduction in \code{tile_obj}
to use for sub-clustering. If NULL, compute PCA on the subsetted cell-level data.}

\item{smooth_emb}{Number of smoothing iterations to perform on the cell embeddings,
as for \code{GetTiles}. If a vector, then embeddings after each specified
iteration are concatenated. If \code{0} is included, then the original embeddings are also included.}

\item{graph.name.cells}{Name of the graph in \code{cell_obj} containing the cell adjacency graph.}

\item{meta.vars.include}{Metadata variables to include in the sub-clustered object.}

\item{harmony.group.by.vars}{Metadata variables to use for Harmony integration.}

\item{early_stop}{Whether to use early stopping in Harmony.}
}
\value{
A Seurat object for the tile subset that can be used for sub-clustering.
}
\description{
Create a Seurat object for sub-clustering a specific cluster from matching tile and cell data
}
