% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_aggregation.R
\name{merge_aggs}
\alias{merge_aggs}
\title{Merges tiles using single-linkage agglomerative clustering}
\usage{
merge_aggs(
  aggs,
  agg_mode,
  d_mu = NULL,
  d_sig = NULL,
  iter_max = NULL,
  dscore_thresh = 0,
  min_npts = 0,
  max_npts = Inf,
  min_area = 0,
  max_area = Inf
)
}
\value{
The \code{aggs} data structure with updated values for
\item{meta_data}{A data table with attributes for the new merged tiles:
* \verb{X,Y}: Centroid of each tile. (NOT updated)
* \code{npts}: Number of points in each tile. (Updated)
* \code{shape}: A \code{sfc} list-column with the geometries for each tile. (NOT updated)
* \verb{area,perimeter}: Area and perimeter of each tile. (Updated)}
\item{edges}{A data table with attributes for edges between adjacent merged tiles,
where edges with infinite \code{dscore} are removed:
* \verb{from,to}: Tile IDs for the two tiles bordering this edge. (Updated for new merged tiles)
* \verb{x0,y0,x1,y1}: Centroid coordinates for the two tiles bordering this edge. (NOT updated)
* \verb{area,npts}: Sum of areas and numbers of points in the two tiles bordering this edge. (Updated)
* \code{edge_length}: Total length of the border between the \code{from} and \code{to} tiles. (Updated)
* \code{dscore}: Overall score for merging two tiles. Product of \code{w}, \code{score_size}, and \code{dC}. (Updated)
* \code{w}: Gene expression similarity score. (Updated)
* \code{score_size}: Penalizes tiles with many points. (Updated)
* \code{perimeter_merge}: Perimeter of merged tile. (Updated)}
\item{pcs}{A \code{num_tiles} x \code{npcs} matrix with the average embedding value over all cells
in each tile. (Updated)}
\item{pcs_merged}{A \code{num_edges} x \code{npcs} matrix with average PCs for merged tiles. (Updated)}
\item{d_mu,d_sig}{Parameters used to calculate \code{w}.}
\item{aggmap}{A length \code{orig_num_tiles} vector mapping each original tile ID to the new
tile IDs after merging.}
}
\description{
Note that this function mutates many of the input values to keep
track of updated values for tiles and their borders after
each successive merge. Every merge of adjacent tiles has an
associated score (higher means merging is more favorable).
Merging is conducted greedily, one step of a time, updating
the score associated with pair of tiles at each step.
}
