% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_analysis.R
\name{MergeClusters}
\alias{MergeClusters}
\title{Merge clusters in a Seurat object}
\usage{
MergeClusters(
  obj,
  to_merge,
  new_label,
  clusters.name = "seurat_clusters",
  new.clusters.name = NULL
)
}
\arguments{
\item{obj}{A Seurat object.}

\item{to_merge}{A vector of cluster labels to merge.}

\item{new_label}{The new label for the merged cluster.}

\item{clusters.name}{Name of the metadata column in \code{obj} containing the cluster labels
to be merged. Defaults to 'seurat_clusters'.}

\item{new.clusters.name}{Name of the metadata column to store the new cluster labels.
If NULL, defaults to \verb{<clusters.name>_m<new_label>}.}
}
\value{
The input Seurat object with the merged cluster labels added to metadata.
}
\description{
Merge clusters in a Seurat object
}
