% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{mergeListsToArmaUVec}
\alias{mergeListsToArmaUVec}
\title{Copies elements from two lists an Armadillo uvec}
\usage{
mergeListsToArmaUVec(list1, list2)
}
\arguments{
\item{list1, list2}{Lists of unsigned values.}
}
\value{
An Armadillo uvec that concatenates \code{list1} and \code{list2}.
}
\description{
Concatenates the values from \code{list1} and \code{list2}.
}
