% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/models.R
\name{neg_hessian_beta}
\alias{neg_hessian_beta}
\title{Compute the Negative Hessian for beta.}
\usage{
neg_hessian_beta(Q_list, tau2_list, X_list)
}
\arguments{
\item{Q_list}{List of the unscaled precision matrices.
One matrix per area, same ordering and length as eta_list.
Depends on gamma, so that dependency is implicit.
I.e., Q_list should be updated in each iteration after gamma is updated.}

\item{tau2_list}{List or vector of precision matrix scaling values.
Same length and ordering as eta_list, with one number per area.}

\item{X_list}{List of covariate matrices.
One matrix per area, same ordering and length as eta_list.}
}
\value{
Negative Hessian matrix of beta.
Also called the observed information matrix.
}
\description{
Compute the Negative Hessian for beta.
}
\note{
Applies to ALL areas.

Requires the Matrix library.
}
\author{
Florica J Constantine, florica AT berkeley.edu
}
