% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/models.R
\name{neg_hessian_gamma_CAR}
\alias{neg_hessian_gamma_CAR}
\title{Compute the Negative Hessian of gamma.
ONLY APPLIES TO THE CAR MODEL.}
\usage{
neg_hessian_gamma_CAR(gamma_hat, eig_vals)
}
\arguments{
\item{gamma_hat}{Parameter of interest: Estimate of correlation parameter.}

\item{eig_vals}{Eigenvalues of Z = D^{-1} W.}
}
\value{
Negative Hessian of gamma.
Also called the observed information.
}
\description{
Compute the Negative Hessian of gamma.
ONLY APPLIES TO THE CAR MODEL.
}
\note{
Applies to a single area.
}
\author{
Florica J Constantine, florica AT berkeley.edu
}
