% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/models.R
\name{neg_hessian_tau2}
\alias{neg_hessian_tau2}
\title{Compute the Negative Hessian for tau^2.}
\usage{
neg_hessian_tau2(Vhat, eta_hat, Q, tau2_hat, beta_hat, X)
}
\arguments{
\item{Vhat}{Estimated covariance matrix of eta.}

\item{eta_hat}{Estimated mean of eta.}

\item{Q}{Unscaled precision matrix.}

\item{tau2_hat}{Parameter of interest: Current estimate of scale.}

\item{beta_hat}{Current estimate of covariate effects.}

\item{X}{Covariate matrix.}
}
\value{
Negative Hessian of tau^2.
Also called the observed information.
}
\description{
Compute the Negative Hessian for tau^2.
}
\note{
Applies to a single area.
}
\author{
Florica J Constantine, florica AT berkeley.edu
}
