% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{normalize}
\alias{normalize}
\title{Helper function for Moran's I: Center a vector.}
\usage{
normalize(x)
}
\arguments{
\item{x}{Measurements.}
}
\value{
x - mean(x).
}
\description{
Helper function for Moran's I: Center a vector.
}
\author{
Matthew Cooper.
}
