% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_pca.R
\name{normalize_data}
\alias{normalize_data}
\title{Log-normalization for counts data}
\usage{
normalize_data(A, scaling_factor = NULL)
}
\arguments{
\item{A}{A genes x cells counts matrix.
Must be convertible to dgCMatrix.}

\item{scaling_factor}{Number of total counts to normalize
all cells to. If NULL, then the median counts across
all cells is used. Defaults to NULL.}
}
\value{
Normalized genes x cells matrix in
dgCMatrix sparse matrix format.
}
\description{
Normalizes all cells to have the same total counts
and computes log(1+x) transform.
}
