% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/poisECM_spNNGP.R
\name{poisECM_spNNGP}
\alias{poisECM_spNNGP}
\title{Fit multi-area Poisson spatial generalized linear model.
Allows for sparse nearest neighbor Gaussian process random effects.
Exponential, Gaussian, Matern, and Spherical covariance structures
are supported.
Fits a common set of fixed effects across areas.}
\usage{
poisECM_spNNGP(
  poisECMData_obj,
  gene_name,
  cov_type = "Exp",
  nngp_k = 20,
  em_iters = 200,
  opt_iters = 10,
  em_min_iters = 10,
  em_tol = 0.001,
  em_stopping = NULL,
  beta_init = "glm",
  cov_init = "BRISC",
  cov_fit_method = "BRISC",
  verbose = FALSE,
  dense_matrices = TRUE
)
}
\arguments{
\item{poisECMData_obj}{Object containing data.
Created by the prepData method.}

\item{gene_name}{Which gene/measurement to fit.
I.e., Which row in the count data matrix to fit.}

\item{cov_type}{Which model to fit for the Gaussian process covariance.
"Exp", "Mat", "Gau", and "Sph" are the valid options, for
Exponential, Matern, Gaussian, and Spherical, respectively.}

\item{nngp_k}{How many neighbors to use for the spNNGP kernel.}

\item{em_iters}{Number of E(C)M iterations to run.}

\item{opt_iters}{Number of inner M-step iterations to run.
The M-step is technically a CM step, i.e., a conditional maximization.
That is, we perform block coordinate ascent wherein we maximize the
expected likelihood in tau^2 and then in gamma for each area, and then
in beta, where each maximization holds all other parameters constant.
This approach leads to a slower convergence in terms of the number of
iterations (em_iters) but faster computation in each iteration.}

\item{em_min_iters}{Minimum number of E(C)M iterations to run.}

\item{em_tol}{Tolerance for early stopping the E(C)M algorithm.}

\item{em_stopping}{How to decide to stop early.
NULL: Don't stop early.
"abs_loglike": Difference in absolute data log likelihood, |old - new|.
"rel_loglike": Relative difference in data log likelihood, |old - new|/|old|.
Data log likelihood taken across alll areas.
"abs_beta_norm": Absolute L2 norm of difference in beta, ||old - new||_2.
"rel_beta_norm": Relative L2 norm of difference in beta,
||old - new||_2 / ||old||_2.}

\item{beta_init}{How to initialize beta.
"glm": Stack up all the observed counts z and covariate matrices X and
fit a Poisson GLM z ~ 0 + X using base R's glm function.
"random": Initialize with standard normal random variables.
A vector or matrix: Initialize with provided value.}

\item{cov_init}{How to initialize covariance parameters.
"BRISC": Call BRISC on log(0.5 + z) - X beta_init.
"variogram": log(0.5 + z) - X beta_init.
A vector of matrix: Initialize with provided value.}

\item{cov_fit_method}{"BRISC" or "variogram": How to perform the M-step
estimation/update of covariance parameters.}

\item{verbose}{Boolean of whether to print out updates.}

\item{dense_matrices}{Boolean of whether to treat Q as a dense matrix when
computing certain quantities in the E-step. This will increase memory usage
(e.g., ~2 -> 8 GB for 3k cells per sample and 20 neighbors), but will dramatically
speed up computation (around a 60-80\% decrease).}
}
\value{
A list comprised of the following:

beta_hat: Estimated coefficients.

cov_param_hat: Estimated spatial covariance parameters.

predictions: Predicted values.
Vector of all values (corresponds to z_list stacked together).

eta_hat: Estimated random effects.
Vector of all values (corresponds to z_list stacked together).

theta_hat: Estimated Poisson parameters.
Vector of all values (corresponds to z_list stacked together).

phi_hat: Estimated spatial random effects.
Vector of all values (corresponds to z_list stacked together).

cov_param_tracker: Estimated correlation parameters.
areas x EM iterations x (Nugget, Sill, Range, Smoothness)---history across iterations.

beta_tracker: Estimated fixed effects.
coefficients x EM iterations matrix---history across iterations.

R2_tracker: Squared correlation between predictions and observations.
areas x EM Iterations matrix---history across iterations.

MSE_tracker: Mean-Square Error between predictions and observations.
areas x EM Iterations matrix---history across iterations.

data_log_like_tracker: Log likelihood of observations.
areas x EM Iterations matrix---history across iterations.

expected_log_like_tracker: Expected log likelihood at current parameters.
areas x EM Iterations matrix---history across iterations.

resid_moran_nb: Moran's I for each area computed using neighbor
adjacency as weights.
areas x EM Iterations matrix---history across iterations.

resid_moran: Moran's I for each area computed using coordinates.
areas x EM Iterations x (value, expectation, sd) array.
NULL unless coordinates are supplied.

start_idx_list: Indices where each area's values start in predictions, etc.
E.g., if area 1 has 100 points and area 2 has 50, we would have
c(1, 101, 151, ...).

time: Total time taken by function.

beta_neghessian: Negative Hessian of final value of beta.
Matrix.

run_settings: A list with the parameter settings used to run the algorithm.

performanceSummary: A dataframe with summary statistics for each sample.
See \code{\link[=summarizePoisECMPerformance]{summarizePoisECMPerformance()}} for more details.
}
\description{
Fit multi-area Poisson spatial generalized linear model.
Allows for sparse nearest neighbor Gaussian process random effects.
Exponential, Gaussian, Matern, and Spherical covariance structures
are supported.
Fits a common set of fixed effects across areas.
}
\note{
Requires the Matrix library.
}
\references{
Meng, Xiao-Li, and Donald B. Rubin.
"Maximum likelihood estimation via the ECM algorithm: A general framework."
Biometrika 80.2 (1993): 267-278.
}
\author{
Florica J Constantine, florica AT berkeley.edu
}
