% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{poisson_loglike}
\alias{poisson_loglike}
\title{Compute the log likelihood of a set of Poisson variables.}
\usage{
poisson_loglike(z, theta_hat)
}
\arguments{
\item{z}{Observed counts.}

\item{theta_hat}{Estimated Poisson parameters.}
}
\value{
log likelihood of data.
}
\description{
Compute the log likelihood of a set of Poisson variables.
}
\note{
Discards invalid values in computation.

Discards factorial terms for numerical stability, etc.
}
\author{
Florica J Constantine, florica AT berkeley.edu
}
