% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepData.R
\name{prepDataSpatialExperiment}
\alias{prepDataSpatialExperiment}
\title{Prepare data for the TESSERA method from a SpatialExperiment object.}
\usage{
prepDataSpatialExperiment(
  spDataObject,
  sample_col,
  design_formula = NULL,
  design_mat = NULL,
  model_type = "Leroux",
  D_THRESH = NULL,
  expected_num_neighbors = 6
)
}
\arguments{
\item{spDataObject}{A SpatialExperiment object.}

\item{sample_col}{String: Column name in the meta data identifying which rows correspond
to which sample (how to break up the data into multiple samples).}

\item{design_formula}{Formula object (compatible with stats::model.matrix).
Supply one of design_formula and design_mat.}

\item{design_mat}{Design matrix.
Supply one of design_formula and design_mat.
Zero columns will be dropped.}

\item{model_type}{One of "Leroux", "CAR", "SAR", "spNNGP".
Spatial covariance structure.}

\item{D_THRESH}{Distance threshold for assigning two observations as adjacent.
Needed to compute adjacency matrices.
Not needed for non-lattice models (set to NULL).}

\item{expected_num_neighbors}{In the absence of a known D_THRESH, how many
neighbors (on average) each cell should have. E.g., Visium data has 6.
Supply at least one of D_THRESH and expected_num_neighbors; D_THRESH supercedes
expected_num_neighbors.}
}
\value{
A list comprised of the following.

coords_list: A list with a matrix of coordinates for each sample.

covariates_list: A list with a dataframe of metadata for each sample.

library_size_list: A list with a vector of library sizes for each sample.
Total counts for each location if there are more than one gene/measurement;
otherwise, set to 1.

X_list: A list with a matrix of design matrices for each sample.

W_list: A list with measurement adjacency matrix for each sample.
If computed.

D_list: A list with measurement degree matrix for each sample.
If computed.

eig_CS_list: A list with CAR/SAR model eigenvalues for each sample.
If computed.

eig_L_list: A list with Leroux model eigenvalues for each sample.
If computed.
}
\description{
Prepare data for the TESSERA method from a SpatialExperiment object.
}
\note{
Calls prepData and visualizeNeighborDistances.

This function is optional/not a core function of the package.
It requires installation of the following Bioconductor packages:
"SpatialExperiment", "SingleCellExperiment", "SummarizedExperiment".
}
\author{
Florica J Constantine, florica AT berkeley.edu
}
