% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{procrustes_inner}
\alias{procrustes_inner}
\title{Computes the procrustes inner product between two matrices}
\usage{
procrustes_inner(A, B)
}
\arguments{
\item{A, B}{Input matrices with same dimensions.}
}
\value{
The procrustes inner product <R*A, B> between A and B.
}
\description{
Solves the orthogonal procrustes problem
R = argmin ||R\emph{A - B||_F
where R is an orthonormal matrix.
Then computes the frobenius inner product
<R}A, B> = <R, B*A^T>,
which is maximized by R.
}
