% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{procrustes_mat}
\alias{procrustes_mat}
\title{Solves the orthogonal procrustes problem}
\usage{
procrustes_mat(A, B)
}
\arguments{
\item{A, B}{Input matrices with same dimensions.}
}
\value{
The orthonormal matrix R that best maps A to B.
}
\description{
Solves the orthogonal procrustes problem
R = argmin ||R*A - B||_F
where R is an orthonormal matrix.
}
