% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_initdata.R
\name{prune_graph}
\alias{prune_graph}
\title{Prunes mesh by eliminating long edges and small connected components}
\usage{
prune_graph(data, thresh_quantile = 0.95, mincells = 10, thresh = NA)
}
\arguments{
\item{data}{A list containing the mesh data structures:
\itemize{
  \item{`pts`}{A V-by-M data table with columns `X` and `Y` containing the
  coordinates of cells and additional metadata.}
  \item{`tris`}{A F-by-4 data table containing the X,Y coordinates of each
   triangle's centroid in the first two columns, and area and
   largest height of each triangle in the last two columns.}
  \item{`edges`}{A E-by-14 data table with columns `from_pt`, `to_pt`, `from_tri`, `to_tri`,
  `x0_pt`, `x1_pt`, `y0_pt`, `y1_pt`, `x0_tri`, `x1_tri`, `y0_tri`, `y1_tri`,
  `length_pt`, `length_tri`. If only one triangle uses an edge, then the `from_tri`,
  `x0_tri`, and `y0_tri` fields will contain NaN values.}
  \item{`tri_to_pt`}{A F-by-V sparse matrix with value 1 at (i,j) if
  triangle i uses point j as a vertex.}
}}

\item{thresh_quantile}{Floating point value between 0 and 1, inclusive.
Quantile of edge length above which edges are pruned. Defaults to 0.95.}

\item{mincells}{Minimum number of cells required for a connected
component of triangles to be kept. Defaults to 10.}
}
\value{
A list containing the mesh data structures with (possibly)
  fewer points, edges, and triangles. Indices have been updated since
  some objects might have been removed.
}
\description{
Pruning works in three steps:
1. Any triangle with an edge longer than a threshold length is removed.
   Afterwards, any edge that no longer belongs to a triangle is removed.
   Then any point that no longer belongs to an edge is removed.
2. Connected components of triangles with a shared edge are computed.
   If everything is connected, then nothing is pruned.
3. Otherwise, components (and corresponding triangles) that contain less
   than `mincells` points are removed. Afterwards, any edge that no
   longer belongs to a triangle is removed. Then any point that no longer
   belongs to an edge is removed.
}
