% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/models.R
\name{Q_matrix_CAR}
\alias{Q_matrix_CAR}
\title{Compute the unscaled precision matrix in a CAR model.}
\usage{
Q_matrix_CAR(W, D, gamma_val)
}
\arguments{
\item{W}{Neighbor/adjacency matrix (symmetric, binary).}

\item{D}{Degree matrix (diagonal, values are row-sums of W).}

\item{gamma_val}{Correlation parameter.}
}
\value{
Unscaled precision matrix.
}
\description{
Compute the unscaled precision matrix in a CAR model.
}
\note{
Applies to a single area.
}
\author{
Florica J Constantine, florica AT berkeley.edu
}
