% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/models.R
\name{Q_matrix_Leroux}
\alias{Q_matrix_Leroux}
\title{Compute the unscaled precision matrix in a Leroux model.}
\usage{
Q_matrix_Leroux(W, D, gamma_val, precomp = NULL)
}
\arguments{
\item{W}{Neighbor/adjacency matrix (symmetric, binary).}

\item{D}{Degree matrix (diagonal, values are row-sums of W).}

\item{gamma_val}{Correlation parameter.}

\item{precomp}{Precomputed list containing D_minus_W and id_mat.}
}
\value{
Unscaled precision matrix.
}
\description{
Compute the unscaled precision matrix in a Leroux model.
}
\note{
Applies to a single area.

Requires the Matrix library.
}
\author{
Florica J Constantine, florica AT berkeley.edu
}
