% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_analysis.R
\name{RunUMAPCustom}
\alias{RunUMAPCustom}
\title{Run UMAP and save fgraph and embeddings in Seurat object}
\usage{
RunUMAPCustom(
  obj,
  reduction = "pca",
  dims = NULL,
  fgraph_only = FALSE,
  graph.name = NULL,
  reduction.name = "umap",
  assay = NULL,
  key = "UMAP_",
  n_neighbors = 30,
  n_components = 2,
  metric = "cosine",
  spread = 1,
  min_dist = 0.3,
  n_threads = NULL,
  fast_sgd = TRUE,
  verbose = TRUE,
  ...
)
}
\arguments{
\item{obj}{A Seurat object.}

\item{reduction}{Name of dimensional reduction in \code{obj} to use as input to UMAP.}

\item{dims}{Dimensions of \code{reduction} to use as input to UMAP. If NULL, use all dimensions.}

\item{fgraph_only}{If TRUE, only compute and store the fuzzy simplicial graph (fgraph) and skip UMAP embedding.}

\item{graph.name}{Name of graph to store the fgraph in \code{obj}. Defaults to \verb{<assay>_fgraph}.}

\item{reduction.name}{Name of dimensional reduction to store the UMAP embeddings in \code{obj}. Defaults to "umap".}

\item{assay}{Assay to set as default assay for the new dimensional reduction. If NULL, use the default assay of \code{obj}.}

\item{key}{Key prefix to use for the new dimensional reduction. Defaults to "UMAP_".}

\item{n_neighbors}{Number of nearest neighbors to use in UMAP. See \code{?uwot::umap} for details.}

\item{n_components}{Number of UMAP dimensions to compute. Ignored if \code{fgraph_only} is TRUE.}

\item{metric}{Distance metric to use in UMAP. See \code{?uwot::umap} for details.}

\item{spread}{UMAP spread parameter. See \code{?uwot::umap} for details.}

\item{min_dist}{UMAP minimum distance parameter. See \code{?uwot::umap} for details.}

\item{n_threads}{Number of threads to use in UMAP. See \code{?uwot::umap} for details.}

\item{fast_sgd}{Whether to use the fast stochastic gradient descent optimization in UMAP. See \code{?uwot::umap} for details.}

\item{verbose}{Whether to print progress messages.}

\item{...}{Additional parameters to pass to \code{uwot::umap}.}
}
\value{
The input Seurat object with the fgraph and (optionally) UMAP embeddings added.
}
\description{
Run UMAP and save fgraph and embeddings in Seurat object
}
