% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_pca.R
\name{scale_data}
\alias{scale_data}
\title{Z-score a sparse matrix across each row or column}
\usage{
scale_data(A, margin = 1, thresh = 10)
}
\arguments{
\item{A}{A log-transformed genes x cells counts matrix.
Must be convertible to dgCMatrix.}

\item{margin}{Subscript over which to compute Z-score.
If \code{1}, then each row is Z-scored; otherwise, each
column is Z-scored.}
}
\value{
Z-scored genes x cells matrix in
dgCMatrix sparse matrix format.
}
\description{
Z-score a sparse matrix across each row or column
}
