% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{scaleRows_dgc}
\alias{scaleRows_dgc}
\title{Z-score a sparse matrix across each row}
\usage{
scaleRows_dgc(x, p, i, ncol, nrow, thresh)
}
\arguments{
\item{x, p, i}{Internal data structures from a
sparse matrix in dgCMatrix format.}

\item{ncol, nrow}{Dimensions of sparse matrix input.}

\item{thresh}{Z-scores above \code{thresh} and below
\code{-thresh} are clipped to \code{thresh} and \code{-thresh},
respectively.}
}
\value{
A dense matrix in column-major ordering
with dimensions \code{nrow} x \code{ncol}.
}
\description{
Z-score a sparse matrix across each row
}
