% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/post_analysis.R
\name{selectWaldStatisticThreshold}
\alias{selectWaldStatisticThreshold}
\title{Given Wald statistics, compute an optimal threshold.}
\usage{
selectWaldStatisticThreshold(
  wald_stats,
  quantile_spacing = 0.01,
  metric = "Raw_MSE"
)
}
\arguments{
\item{wald_stats}{Vector of Wald statistics: non-negative F-statistics.
Square t-statistics to obtain F-statistics.}

\item{quantile_spacing}{How finely to search for a threshold.
Default is 0.01, for 0.01, 0.02, ..., 0.99 quantiles.}

\item{metric}{What metric to use for selecting the optimal threshold.
Options are "Raw_" or "Log_" plus "MSE", "MAE", "MedAE", or "MaxAE",
for the Mean-Square Error (L2), the Mean Absolute Error (L1), the Median
Absolute Error, and the Maximum Absolute Error (L-infinity/KS).}
}
\value{
A list with the following entries

threshold The optimal threshold (lowest error).

chi2params A vector of the shift and scale parameters corresponding
to the optimal threshold.

threshold_results A dataframe with all metrics and fitted parameters
for all thresholds/quantiles.
}
\description{
In the TESSERA algorithm, and in generalized linear mixed models in general,
the exact distribution of p-values for Wald statistics under the null hypothesis
is unknown.
We know that the standard errors are, for finite samples, likely biased
downward (too small).
Moreover, for finite samples, due to the approximations
in the E-step of the fitting algorithm, there may be a small positive bias.
Hence, a scaled chi-squared distribution or scaled F-distribution with
a non-centrality parameter might be a good model for the statistics under
the null hypothesis, where we note that an unscaled chi-squared or F distribution
are the standard asymptotic distributions for Wald statistics.
To use this function, the Wald statistics as well as a threshold is estimated.
The threshold is a rough upper bound for the statistics coming from the null
distribution--it need not be exact, but it should be greater than most of the
null distributed statistics and smaller than most of the non-null distributed
statistics.
The threshold is estimated by noting that under the null hypothesis,
the p-values should be approximately Uniform(0, 1)
distributed; for each threshold value, we compute the error
between the quantiles of the p-values and those of the Uniform(0, 1) distribution
and choose the threshold with the lowest error.
}
\examples{
selectWaldStatisticThreshold(c(5 * stats::rchisq(2000, 1, ncp = 10),
 200 + 5 * stats::rchisq(2000, 1, ncp = 10)), 0.05)
}
\author{
Florica J Constantine, florica AT berkeley.edu
}
