% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tessera.R
\name{set.aneuploid}
\alias{set.aneuploid}
\title{Set a cell to contain an aneuploid chromosome}
\usage{
set.aneuploid(embryo, cell.index, chromosome)
}
\arguments{
\item{embryo}{the embryo}

\item{cell.index}{the cell to affect}

\item{chromosome}{the chromosome to make aneuploid}
}
\value{
the modified embryo
}
\description{
Set a cell to contain an aneuploid chromosome
}
