% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tessera.R
\name{set.aneuploidies}
\alias{set.aneuploidies}
\title{Set aneuploidies in an embryo}
\usage{
set.aneuploidies(embryo, chromosome, prop.aneuploid, dispersion)
}
\arguments{
\item{embryo}{an embryo as created by \code{create.embryo}}

\item{chromosome}{the chromosome to set aneuploidies for}

\item{prop.aneuploid}{the proportion of aneuploid cells (0-1)}

\item{dispersion}{the dispersion of the aneuploid cells (0-1)}
}
\value{
the embryo with aneuploidies
}
\description{
A sphere of cells is created with the given proportion of aneuploidies.
Aneuploid cells are either adjacent or dispersed
}
\examples{
embryo <- set.aneuploidies(embryo, 1, 0.1, 0.9)
}
