% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_pca.R
\name{smooth_embedding}
\alias{smooth_embedding}
\title{Smooth embeddings along edges}
\usage{
smooth_embedding(dmt, smooth_emb = 0)
}
\arguments{
\item{dmt}{A DMT object.}

\item{smooth_emb}{Number of smoothing iterations to perform.
If a vector, then embeddings after each specified
iteration are concatenated. If \code{0} is included, then
the original embeddings are also included.}
}
\value{
The input DMT object with smoothed embeddings stored in
\code{dmt$udv_cells$embeddings}.
}
\description{
Smooth embeddings along edges
}
