% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_gradients_edges.R
\name{smooth_field_edges}
\alias{smooth_field_edges}
\title{Bilateral / anisotropic filtering of gradient field}
\usage{
smooth_field_edges(
  edges,
  field,
  edges_svd,
  coords,
  adj_idx,
  distance = "projected",
  similarity = "euclidean"
)
}
\description{
Bilateral / anisotropic filtering of gradient field
}
