% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{sparseDist_LT}
\alias{sparseDist_LT}
\title{Get k nearest neighbors/distances given coordinates.
Modified from: https://stackoverflow.com/questions/5560218/computing-sparse-pairwise-distance-matrix-in-r}
\usage{
sparseDist_LT(coords, k)
}
\arguments{
\item{coords}{(x, y) for points as rows.}

\item{k}{Number of nearest neighbors.}
}
\value{
2 k x n points matrix.
First k rows are distances, last k are indices of nearest neighbors.
Only Lower Triangular part is formed.
}
\description{
Get k nearest neighbors/distances given coordinates.
Modified from: https://stackoverflow.com/questions/5560218/computing-sparse-pairwise-distance-matrix-in-r
}
