% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tessera.R
\name{take.one.biopsy}
\alias{take.one.biopsy}
\title{Take a sample from an embryo}
\usage{
take.one.biopsy(embryo, n.sampled.cells, index.cell)
}
\arguments{
\item{embryo}{an embryo as created by `create.embryo`}

\item{n.sampled.cells}{the number of cells to biopsy}

\item{index.cell}{the index of the cell to begin biopsying. Must be a value
between 1 and `nrow(embryo)`}
}
\value{
the number of aneuploid cells in the biopsy
}
\description{
The cell at the given index is taken,
plus the closest n neighbouring cells where n = n.sampled.cells-1.
}
\examples{
e <- create.embryo(100, 0.1, 0.1)
take.one.biopsy(e, 5, 1)
}
