% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tessera-package.R
\docType{package}
\name{tessera}
\alias{tessera-package}
\alias{tessera}
\title{tessera}
\description{
Spatial tiling
}
\author{
\strong{Maintainer}: Ilya Korsunsky \email{ilya.korsunsky@gmail.com} (\href{https://orcid.org/0000-0003-4848-3948}{ORCID})

Authors:
\itemize{
  \item Daniel Stein \email{daniel_stein@g.harvard.edu} (\href{https://orcid.org/0000-0002-9442-2604}{ORCID})
}

}
