% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\name{tessera_example_data}
\alias{tessera_example_data}
\title{Synthetic TESSERA Example Data Files}
\description{
The following files are stored in \code{inst/extdata} and are
used for runnable examples and unit tests within the TESSERA package.
}
\section{Data Generation}{

All datasets below were generated using the script located at:
\code{system.file("examples", "gen_data_and_run.R", package = "TESSERA")}
}

