% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{tessera_warmup}
\alias{tessera_warmup}
\title{Sample data for Tessera vignettes}
\format{
counts and meta data table for 3177 cells.
\describe{
\item{counts}{dgCMatrix of gene-by-cell counts}
\item{meta_data}{data.frame with spatial locations (XY) and cell lineage (type)}
}
}
\source{
Chen, J.H., Nieman, L.T., Spurrell, M. et al. Human lung cancer harbors spatially organized stem-immunity hubs associated with response to immunotherapy. Nat Immunol 25, 644–658 (2024). https://doi.org/10.1038/s41590-024-01792-2
}
\usage{
tessera_warmup
}
\description{
Subset of MERFISH lung cancer dataset (Chen et al 2024).
}
\keyword{datasets}
