% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{trace_back_cpp}
\alias{trace_back_cpp}
\title{Trace back from a point to its root in the spanning forest}
\usage{
trace_back_cpp(v0, vcrit, parent_edge, parent)
}
\arguments{
\item{v0}{Index of starting point (0-indexed).}

\item{vcrit}{Index of critical point associated with the tree that \code{v0} belongs to (0-indexed).}

\item{parent_edge}{A length \code{num_points} vector containing the directed edge that has
each point as a target node. Critical points have no parent edge, so the value is ignored.}

\item{parent}{A length \code{num_points} vector containing the parent (source) point for each
point in the directed spanning forest. Critical points have no parent, so the value is ignored.}
}
\value{
Vector of edge indices along the path from \code{v0} to its root \code{vcrit} in the tree (1-indexed).
}
\description{
Trace back from a point to its root in the spanning forest
}
