% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{trace_epaths_cpp}
\alias{trace_epaths_cpp}
\title{Trace all paths from saddles to critical points in the spanning forest}
\usage{
trace_epaths_cpp(saddles, vcrits, edges_from, edges_to, parent_edge, parent)
}
\arguments{
\item{saddles}{A length \code{num_saddles} vector with edge indices for saddle edges (0-indexed).}

\item{vcrits}{A length \code{num_critpts} vector of critical points.}

\item{edges_from}{A length \code{num_edges} vector with indices for the first end-point of each edge
in the mesh (0-indexed).}

\item{edges_to}{A length \code{num_edges} vector with indices for the second end-point of each edge
in the mesh (0-indexed).}

\item{parent_edge}{A length \code{num_points} vector containing the directed edge that has
each point as a target node in the directed spanning forest. Critical points have
no parent edge, so the value is ignored.}

\item{parent}{A length \code{num_points} vector containing the parent (source) point for each
point in the directed spanning forest. Critical points have no parent, so the value is ignored.}
}
\value{
A list of length \code{2*num_saddles} containing the two paths from each saddle edge to the
two critical points that it joins. Each path is a numeric vector of edge indices (1-indexed).
}
\description{
Trace all paths from saddles to critical points in the spanning forest
}
