% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_shapes.R
\name{trace_polygons}
\alias{trace_polygons}
\title{Contruct shapes that outline each tile}
\usage{
trace_polygons(dmt, aggs)
}
\arguments{
\item{dmt}{A DMT mesh data structure with \code{pts}, \code{edges}, and \code{tris} attributes.
Assignment of points to tiles is provided in \code{dmt$pts$agg_id}.}

\item{aggs}{A tile data structure used to specify the total number of tiles.}
}
\value{
A \code{sfc} list-column with the geometries for each tile.
}
\description{
Contruct shapes that outline each tile
}
