% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{trace_polygons_cpp}
\alias{trace_polygons_cpp}
\title{Contruct shapes that outline each tile}
\usage{
trace_polygons_cpp(edges, naggs, ntris, pts_dmt_component)
}
\arguments{
\item{edges}{A matrix with \code{num_edges} rows with mesh edge information.}

\item{naggs}{Number of tiles.}

\item{pts_dmt_component}{A length \code{num_points} vector with the unique ID for the
tile that each point belongs to.}

\item{ntrix}{Number of triangles in the mesh.}
}
\value{
A list of length \code{naggs} matrices that contain coordinates of the
polygons that trace the outline of each tile.
}
\description{
Contruct shapes that outline each tile
}
