% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_aggregation.R
\name{update_agg_shapes}
\alias{update_agg_shapes}
\title{Update shapes and counts matrix for tiles after merging}
\usage{
update_agg_shapes(dmt, aggs)
}
\arguments{
\item{dmt}{DMT data structure}

\item{agg}{Tile data structure after merging}
}
\value{
\code{aggs} tile data structure with updated values for:
\item{meta_data$shape}{Retraces the polygons for tiles using the new point assignments.}
\item{meta_data$X,meta_data$Y}{Recomputes the centroid of each tile from the polygon shape.}
\item{edges}{Ensures that \code{aggs$edges} always has \code{from} < \code{to}.}
\item{counts}{A \code{num_genes} x \code{num_tiles} gene-by-tile matrix of aggregated transcript counts.}
}
\description{
Update shapes and counts matrix for tiles after merging
}
