% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_aggregation.R
\name{update_dmt_aggid}
\alias{update_dmt_aggid}
\title{Update tile IDs in the DMT data structure after merging}
\usage{
update_dmt_aggid(dmt, aggs)
}
\arguments{
\item{dmt}{DMT data structure}

\item{agg}{Tile data structure after merging, with \code{aggmap} attribute
mapping original tile IDs to new IDs for the merged tiles.}
}
\value{
DMT data structure with updated values for these attributes:
\itemize{
\item \code{pts$agg_id}: Unique ID for the tile that each point belongs to.
\item \code{edges$agg_from}: Unique ID for the tile that \code{edges$from_pt} belongs to.
\item \code{edges$agg_to}: Unique ID for the tile that \code{edges$to_pt} belongs to.
}
}
\description{
Update tile IDs in the DMT data structure after merging
}
