% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{update_E_cpp}
\alias{update_E_cpp}
\title{Updates information for boundaries after merging two tiles}
\usage{
update_E_cpp(
  V_pcs,
  V_perimeter,
  V_area,
  V_npts,
  E_from,
  E_to,
  E_npts,
  E_area,
  E_edge_length,
  E_pcs_merge,
  E_w,
  E_perimeter_merge,
  E_score_size,
  E_dscore,
  e_update,
  V_to_E_from,
  V_to_E_to,
  d_mu,
  d_sig,
  agg_mode,
  min_npts,
  max_npts
)
}
\arguments{
\item{e_update}{Edges that should be updated.}
}
\description{
For every boundary that exists between the new merged tile and other tiles,
the following values must be updated:
\itemize{
\item \code{E_pcs_merge[e_update,]}: PCs that would result from future merging.
\item \code{E_perimeter_merge[e_update]}: Perimeters that would result from future merging.
\item \code{E_w[e_update]}: Expression similarity score.
\item \code{E_score_size[e_update]}: Size of merged tile score.
\item \code{dC[e_update]}: Delta shape compactness score for merged tile.
\item \code{E_dscore[e_update]}: Overall merging score (product of \code{w}, \code{score_size}, \code{dC})
Note that \code{E_npts} and \code{E_area} should already have been previously updated.
}
}
