% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{update_V_cpp}
\alias{update_V_cpp}
\title{Updates information for tiles after merging two tiles}
\usage{
update_V_cpp(
  V_pcs,
  V_npts,
  V_perimeter,
  V_area,
  e_merge_from,
  e_merge_to,
  e_merge_edge_length,
  e_merge_area,
  e_merge_npts,
  e_merge_pcs,
  agg_mode
)
}
\description{
Uses the information from the shared edge to update the merged PCs,
area, number of points, and perimeter. Mutates the \code{e_merge_from}th value
within the \code{V_pcs}, \code{V_npts}, \code{V_perimeter}, \code{V_area} input data structures.
}
