% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{variables_from_list}
\alias{variables_from_list}
\title{Utility function to instantiate variables from elements in a named list.
E.g., call this function on list(a=1, b=2) would result in variables a and b
with values 1 and 2, respectively, in the calling frame/environment.}
\usage{
variables_from_list(lst, target_environ = parent.frame())
}
\arguments{
\item{lst}{List with named fields.}

\item{target_environ}{Environment to instantiate variables in.}
}
\description{
Utility function to instantiate variables from elements in a named list.
E.g., call this function on list(a=1, b=2) would result in variables a and b
with values 1 and 2, respectively, in the calling frame/environment.
}
\note{
This function is not currently used elsewhere in the codebase.

This function has identical functionality and essentially identical code
to extract.named in mvbutils.
}
\author{
Florica J Constantine, florica AT berkeley.edu
}
