% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/comparisons.R
\name{BRISC_wrapper}
\alias{BRISC_wrapper}
\title{Wrapper to fit sparse NN GP model via BRISC.}
\usage{
BRISC_wrapper(
  z_list,
  X_list,
  coords_list,
  k = 15,
  cov_type = "Exp",
  transform_z = TRUE,
  z_offset = 0.5,
  verbose = FALSE
)
}
\arguments{
\item{z_list}{List of vectors of observed counts---one vector per area.}

\item{X_list}{List of design or covariate matrices---one matrix per area.
Same length and ordering as z_list.
Matrices with number of rows equal to length of corresponding vector in
z_list.}

\item{coords_list}{List of coordinate matrices (x, y)---one matrix per area.
Same length and ordering as z_list.
Matrices with number of rows equal to length of corresponding vector in
z_list.}

\item{k}{Number of neighbors.}

\item{cov_type}{String for covariance model type.
"Exp", "Sph", "Gau", and "Mat" are supported.}

\item{transform_z}{Boolean: log-transform z or not.}

\item{z_offset}{If transform_z, the counts z are transformed as
log(z + z_offset).}

\item{verbose}{Whether to print output as BRISC runs.}
}
\value{
The output list from BRISC, plus a time field for how long the
function ran for.
}
\description{
This function theoretically works for multiple areas by stacking everything
together and ignoring differences in coordinates.
}
\note{
Requires the BRISC library.
}
\author{
Florica J Constantine, florica AT berkeley.edu
}
