% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/poisECM_lattice.R
\name{checkInputsPoisECM}
\alias{checkInputsPoisECM}
\title{Check inputs for the poisECM method.
Make sure that the input object has everything needed to run without error.}
\usage{
checkInputsPoisECM(poisECMData_obj)
}
\arguments{
\item{poisECMData_obj}{Object containing data.
Created by the prepData method.}
}
\description{
Check inputs for the poisECM method.
Make sure that the input object has everything needed to run without error.
}
\note{
Does not return anything.

This method can be used to check a hand-created input object.
E.g., if a user does not want to use prepData.
}
\author{
Florica J Constantine, florica AT berkeley.edu
}
