% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tessera.R
\name{create.embryo}
\alias{create.embryo}
\title{Create an embryo}
\usage{
create.embryo(n.cells, prop.aneuploid, dispersion)
}
\arguments{
\item{n.cells}{the number of cells in the embryo}

\item{prop.aneuploid}{the proportion of aneuploid cells (0-1)}

\item{dispersion}{the dispersion of the aneuploid cells (0-1)}
}
\value{
an embryo data frame
}
\description{
A sphere of cells is created with the given proportion of aneuploidies.
Aneuploid cells are either adjacent or dispersed
}
\examples{
embryo <- create.embryo(20, 0.1, 0.9)
}
