% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_dmt.R
\name{dmt_init_tiles}
\alias{dmt_init_tiles}
\title{Initialize tiles with shapes and other properties}
\usage{
dmt_init_tiles(dmt)
}
\arguments{
\item{dmt}{A DMT data structure with \code{pts}, \code{edges}, and \code{udv_cells} attributes.
Assignment of points to tiles is provided in \code{dmt$pts$agg_id}, \code{dmt$edges$agg_from},
and \code{dmt$edges$agg_to}.}
}
\value{
A list data structure for tiles and their adjacencies. Includes the following attributes:
\item{meta_data}{A data table with attributes for tiles:
\item{X,Y}{Centroid of each tile.}
\item{npts}{Number of points in each tile.}
\item{shape}{A \code{sfc} list-column with the geometries for each tile.}
\item{area,perimeter}{Area and perimeter of each tile.}
}
\item{edges}{A data table with attributes for edges between adjacent tiles:
\item{from,to}{Tile IDs for the two tiles bordering this edge.}
\item{x0,y0,x1,y1}{Centroid coordinates for the two tiles bordering this edge.}
\item{area,npts}{Sum of areas and numbers of points in the two tiles bordering this edge.}
\item{edge_length}{Total length of the border between the \code{from} and \code{to} tiles.}
}
\item{pcs}{A \code{num_tiles} x \code{npcs} matrix with the average embedding value over all cells
in each tile.}
}
\description{
Initialize tiles with shapes and other properties
}
