% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tessera.R
\name{.count.empty.blocks}
\alias{.count.empty.blocks}
\title{Count the number of potential seed locations}
\usage{
.count.empty.blocks(d)
}
\arguments{
\item{d}{the embryo}

\item{n.sampled.cells}{the number of cells to biopsy}

\item{seed.sample}{the index of the cell to begin biopsying}
}
\value{
the number of potential seeds
}
\description{
Cells that are not adjacent to a seed
}
