% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/E_step.R
\name{E_step_Vhat}
\alias{E_step_Vhat}
\title{Compute the Covariance Matrix of the random effects eta.
Part of the E-Step in the EM algorithm.
Follows Clayton and Kaldor (1987) for the Poisson-Gaussian model.}
\usage{
E_step_Vhat(Q, tau2, z)
}
\arguments{
\item{Q}{Unscaled precision matrix.}

\item{tau2}{Precision/covariance matrix scaling.}

\item{z}{Observed counts.}
}
\value{
Estimated covariance matrix.
}
\description{
Compute the Covariance Matrix of the random effects eta.
Part of the E-Step in the EM algorithm.
Follows Clayton and Kaldor (1987) for the Poisson-Gaussian model.
}
\note{
Applies to a single area.

This method is an Empirical Bayes approximation.

This method inverts a matrix, because the inverse is actually needed...

Requires the Matrix library.

Requires the pracma library.
}
\author{
Florica J Constantine, florica AT berkeley.edu
}
