% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/models.R
\name{expected_loglike}
\alias{expected_loglike}
\title{Compute the expected log likelihood.}
\usage{
expected_loglike(
  Vhat,
  eta_hat,
  Q,
  gamma_hat,
  tau2,
  beta_hat,
  X,
  W,
  D,
  eig_vals,
  model_type
)
}
\arguments{
\item{Vhat}{Estimated covariance matrix of eta.}

\item{eta_hat}{Estimated mean of eta.}

\item{Q}{Unscaled precision matrix.
The scaled precision matrix in spNNGP.}

\item{gamma_hat}{Parameter of interest: Estimate of correlation parameter.}

\item{tau2}{Parameter of interest: Current estimate of scale.}

\item{beta_hat}{Current estimate of covariate effects.}

\item{X}{Covariate matrix.}

\item{W}{Neighbor/adjacency matrix (symmetric, binary).}

\item{D}{Degree matrix (diagonal, values are row-sums of W).}

\item{eig_vals}{Eigenvalues of Z = D^\{-1\} W (CAR/SAR), D - W (Leroux), or Q (spNNGP).}

\item{model_type}{"CAR", "SAR", or "Leroux", or "spNNGP".
If spNNGP, set tau^2 to 1 and the value of gamma is irrelevant.
Model for the random effects.}
}
\value{
Expected log likelihood.
}
\description{
Compute the expected log likelihood.
}
\note{
Applies to a single area.

Multivariate normal; drops the 2 pi term.
}
\author{
Florica J Constantine, florica AT berkeley.edu
}
