% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/post_analysis.R
\name{fit_scaled_noncentral_chi2}
\alias{fit_scaled_noncentral_chi2}
\title{Fit a scaled non-central chi^2_1 distribution using BOBYQA}
\usage{
fit_scaled_noncentral_chi2(wald_stats, wald_thresh)
}
\arguments{
\item{wald_stats}{Vector of Wald statistics (non-negative).}

\item{wald_thresh}{Threshold below which data is included in the fit.}
}
\value{
Vector of c(scale, shift).
}
\description{
Numerically maximizes the likelihood of a scaled non-central chi-square
distribution (df=1) truncated above a specific threshold.
The fit is conditional on the threshold.
}
\examples{
fit_scaled_noncentral_chi2(5 * stats::rchisq(1000, 1, ncp=10), 20)
}
