% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_gen.R
\name{generate_data_one_area}
\alias{generate_data_one_area}
\title{Generate data for one area.}
\usage{
generate_data_one_area(
  n_points,
  nb_dist,
  model_type,
  beta_true,
  gamma_true = NULL,
  tau2_true = NULL,
  X_type = "rand_bern",
  ar_gamma = 0.75,
  X = NULL,
  library_size = NULL
)
}
\arguments{
\item{n_points}{Number of points in the area.}

\item{nb_dist}{Distance threshold for determining if two points are neighbors.
Coordinates are samples in (0, 1)^2 so 0.03 is a reasonable value.}

\item{model_type}{"CAR", "SAR", or "Leroux"---model for random effects.}

\item{beta_true}{True fixed effects.}

\item{gamma_true}{True correlation parameter.}

\item{tau2_true}{True scale parameter.}

\item{X_type}{How to generate the covariates X.
Note that the coordinates are sorted by x then by y, so nearby points are
more likely to be neighbors.
"rand_bern": X_\{i, j\} ~ 2 ( sign(N(0, 1)) + 1 ) in \{0, 1\}
"rand_unif": X_\{i, j\} ~ U(0, 1)
"rand_norm": X_\{i, j\} ~ N(0, 1)
"ar1": X_\{:, j\} follows an AR(1) process.
"ar1_bern": Generate the AR(1) as above, and apply the same sign transformation
as in "rand_bern"
"intercept": Constant column of all 1s}

\item{ar_gamma}{Autocorrelation parameter for X if needed.}

\item{X}{Instead of passing in X_type, pass in a pre-defined matrix X.}

\item{library_size}{Instead of using all 1s, pass in a pre-specified library size.
Expected value/offset for each point.}
}
\value{
A list with the following fields:

X: Binary covariates.

W: Adjacency matrix.

D: Diagonal degree matrix (row-sums of W).

eig_list: List of eigenvalues of Z = D^\{-1\} W (CAR/SAR) or D - W (Leroux).

library_size: Scaling for theta.

x_coords: x-coordinates of points.

y_coords: y-coordinates of points.

Q: Unscaled precision matrix. Depends on W, D, gamma_true.

phi_true: True sampled random effects.
(multivariate normal with mean zero and covariance tau^2 Q^\{-1\}).

eta_true: phi + X beta.

theta_true: exp(eta).

z: Sampled counts Pois(theta x lib size).
}
\description{
Generate data for one area.
}
\note{
Requires the Matrix library.

Calls the various Q_matrix creation functions.
}
\author{
Florica J Constantine, florica AT berkeley.edu
}
