% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/M_step.R
\name{M_step_beta}
\alias{M_step_beta}
\title{Maximize the expected likelihood in beta, holding other variables constant.
Part of the M-Step in the EM algorithm.}
\usage{
M_step_beta(eta_list, Q_list, tau2_list, X_list)
}
\arguments{
\item{eta_list}{List of the estimated means of eta.
One vector per area.}

\item{Q_list}{List of the unscaled precision matrices.
One matrix per area, same ordering and length as eta_list.
Depends on gamma, so that dependency is implicit.
I.e., Q_list should be updated in each iteration after gamma is updated.}

\item{tau2_list}{List or vector of precision matrix scaling values.
Same length and ordering as eta_list, with one number per area.}

\item{X_list}{List of covariate matrices.
One matrix per area, same ordering and length as eta_list.}
}
\value{
Maximum likelihood estimate of covariates beta.
}
\description{
Maximize the expected likelihood in beta, holding other variables constant.
Part of the M-Step in the EM algorithm.
}
\note{
Applies to ALL areas.

Requires the Matrix library.

Requires the pracma library.
}
\author{
Florica J Constantine, florica AT berkeley.edu
}
