% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/M_step.R
\name{M_step_gamma_CAR}
\alias{M_step_gamma_CAR}
\title{Maximize the expected likelihood in gamma, holding other variables constant.
Part of the M-Step in the EM algorithm.
ONLY APPLIES TO THE CAR MODEL.}
\usage{
M_step_gamma_CAR(
  Vhat,
  eta_hat,
  tau2,
  beta_hat,
  X,
  W,
  D,
  eig_vals,
  gamma_current = NULL
)
}
\arguments{
\item{Vhat}{Estimated covariance matrix of eta.}

\item{eta_hat}{Estimated mean of eta.}

\item{tau2}{Precision matrix scaling parameter tau^2.}

\item{beta_hat}{Current estimate of covariate effects.}

\item{X}{Covariate matrix.}

\item{W}{Neighbor/adjacency matrix (symmetric, binary).}

\item{D}{Diagonal degree matrix (row sums of W).
Not needed, but passed in for consistency with API.}

\item{eig_vals}{Eigenvalues of Z = D^\{-1\} W.}

\item{gamma_current}{Previous value of gamma.
Only needed for failure modes.}
}
\value{
Maximum likelihood estimate of correlation parameter gamma.
A list with gamma_hat (estimate) and grad_val (gradient).
}
\description{
Maximize the expected likelihood in gamma, holding other variables constant.
Part of the M-Step in the EM algorithm.
ONLY APPLIES TO THE CAR MODEL.
}
\note{
Applies to a single area.

Requires the pracma library.
}
\author{
Florica J Constantine, florica AT berkeley.edu
}
