% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/comparisons.R
\name{mgcv_gam_wrapper}
\alias{mgcv_gam_wrapper}
\title{Wrapper to fit a spatial spline model.}
\usage{
mgcv_gam_wrapper(
  z_list,
  X_list,
  x_coords_list,
  y_coords_list,
  model_family = "poisson",
  z_offset = 0.5,
  spline_k = -1,
  spline_basis = "tp"
)
}
\arguments{
\item{z_list}{List of vectors of observed counts---one vector per area.}

\item{X_list}{List of design or covariate matrices---one matrix per area.
Same length and ordering as z_list.
Matrices with number of rows equal to length of corresponding vector in
z_list.}

\item{x_coords_list}{List of x-coordinates for observations---one vector per area.
Same length and ordering as z_list.
Vectors lengths equal to length of corresponding vector in z_list, i.e.,
a 1:1 correspondence.}

\item{y_coords_list}{List of y-coordinates for observations---one vector per area.
Same length and ordering as z_list.
Vectors lengths equal to length of corresponding vector in z_list, i.e.,
a 1:1 correspondence.}

\item{model_family}{"poisson" or "gaussian".
Which model family to fit. Note that Poisson models don't scale well in the
multi-area setting---the spline is fit on a per-area basis.}

\item{z_offset}{In the "gaussian" case, the counts z are transformed as
log(z + z_offset).}

\item{spline_k}{Dimension of spline basis.}

\item{spline_basis}{Type of basis---must be valid for mgcv::s.}
}
\value{
The output list from mgcv::gam, plus a time field for how long the
function ran for.
}
\description{
This function theoretically works for multiple areas.
}
\note{
Requires the mgcv library.
}
\author{
Florica J Constantine, florica AT berkeley.edu
}
