% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{moran_I_nb}
\alias{moran_I_nb}
\title{Compute Moran's I using an adjacency matrix.}
\usage{
moran_I_nb(y, W)
}
\arguments{
\item{y}{A vector of measurements.}

\item{W}{Neighbor/adjacency matrix (symmetric, binary).}
}
\value{
Moran's I (spatial autocorrelation).
}
\description{
Compute Moran's I using an adjacency matrix.
}
\examples{
set.seed(2026)
tau2_true <- 1.0
gamma_true <- 0.5
beta_true <- c(1, 0, -1) 
ex_data <- generate_data_one_area(1000, 0.03, "Leroux", beta_true, 
  gamma_true, tau2_true, "rand_bern")
moran_I_nb(ex_data$z, ex_data$W)
}
\author{
Florica J Constantine, florica AT berkeley.edu
}
