% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{moran_I_nb}
\alias{moran_I_nb}
\title{Compute Moran's I using an adjacency matrix.}
\usage{
moran_I_nb(y, W)
}
\arguments{
\item{y}{A vector of measurements.}

\item{W}{Neighbor/adjacency matrix (symmetric, binary).}
}
\value{
Moran's I (spatial autocorrelation).
}
\description{
Compute Moran's I using an adjacency matrix.
}
\author{
Florica J Constantine, florica AT berkeley.edu
}
