% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/models.R
\name{neg_hessian_gamma_SAR}
\alias{neg_hessian_gamma_SAR}
\title{Compute the Negative Hessian of gamma.
ONLY APPLIES TO THE SAR MODEL.}
\usage{
neg_hessian_gamma_SAR(
  Vhat,
  eta_hat,
  gamma_hat,
  tau2,
  beta_hat,
  X,
  W,
  D,
  eig_vals
)
}
\arguments{
\item{Vhat}{Estimated covariance matrix of eta.}

\item{eta_hat}{Estimated mean of eta.}

\item{gamma_hat}{Parameter of interest: Estimate of correlation parameter.}

\item{tau2}{Parameter of interest: Current estimate of scale.}

\item{beta_hat}{Current estimate of covariate effects.}

\item{X}{Covariate matrix.}

\item{W}{Neighbor/adjacency matrix (symmetric, binary).}

\item{D}{Degree matrix (diagonal, values are row-sums of W).}

\item{eig_vals}{Eigenvalues of Z = D^\{-1\} W.}
}
\value{
Negative Hessian of gamma.
Also called the observed information.
}
\description{
Compute the Negative Hessian of gamma.
ONLY APPLIES TO THE SAR MODEL.
}
\note{
Applies to a single area.

Requires the Matrix library.
}
\author{
Florica J Constantine, florica AT berkeley.edu
}
