% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/models.R
\name{nngp_prec_mat}
\alias{nngp_prec_mat}
\title{Form a sparse Nearest-Neighbor Gaussian Process Precision Matrix.
(Documentation truncated for brevity)}
\usage{
nngp_prec_mat(sp_dist, nb_dist, cov_type, cov_params)
}
\arguments{
\item{sp_dist}{Output of sparse_dist_LT function.}

\item{nb_dist}{Precomputed list of Euclidean distance matrices for neighbors.}

\item{cov_type}{String for covariance model type.}

\item{cov_params}{Covariance/kernel parameters.}
}
\value{
A list with a sparse precision matrix Q, eigenvalues Dinv, and factor A.
}
\description{
Form a sparse Nearest-Neighbor Gaussian Process Precision Matrix.
(Documentation truncated for brevity)
}
