% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/models.R
\name{nngp_prec_mat}
\alias{nngp_prec_mat}
\title{Form a sparse Nearest-Neighbor Gaussian Process Precision Matrix.
See https://mc-stan.org/users/documentation/case-studies/nngp.html.}
\usage{
nngp_prec_mat(sp_dist, coords, cov_type, cov_params)
}
\arguments{
\item{sp_dist}{Output of sparse_dist_LT function.
Basically, top few rows are nearest distances, bottom few rows are indices.
CALL THAT FUNCTION OR SEE IT FOR MORE DETAILS.}

\item{coords}{Matrix (x, y) of coordinates.}

\item{cov_type}{String for covariance model type.
"Exp", "Sph", "Gau", and "Mat" are supported.}

\item{cov_params}{Covariance/kernel parameters.
Same order as in variogram functions.
Nugget variance (partial sill), Spatial variance (partial sill),
Spatial range (scales distance); if Matern, smoothness kappa.}
}
\value{
A list with a sparse precision matrix and associated eigenvalues.

Sparse precision matrix Q.

Eigenvalues Dinv.

Lower triangular factor A.
}
\description{
Form a sparse Nearest-Neighbor Gaussian Process Precision Matrix.
See https://mc-stan.org/users/documentation/case-studies/nngp.html.
}
\note{
Applies to a single area.

Requires the Matrix library.
}
\author{
Florica J Constantine, florica AT berkeley.edu
}
