% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepData.R
\name{prepData}
\alias{prepData}
\alias{prepData,ANY-method}
\alias{prepData,SpatialExperiment-method}
\title{Prepare data for the TESSERA method.}
\usage{
prepData(x, ...)

\S4method{prepData}{ANY}(
  x,
  meta_data,
  sample_col = "sample_id",
  design_formula = NULL,
  design_mat = NULL,
  coord_data = NULL,
  D_THRESH = NULL,
  expected_num_neighbors = 6,
  k_search = 20,
  adj_mat = NULL,
  model_type = "Leroux"
)

\S4method{prepData}{SpatialExperiment}(
  x,
  sample_col = "sample_id",
  design_formula = NULL,
  design_mat = NULL,
  model_type = "Leroux",
  D_THRESH = NULL,
  expected_num_neighbors = 6,
  k_search = 20,
  adj_mat = NULL
)
}
\arguments{
\item{x}{Variables x measurements (genes x cells) count matrix (can be sparse),
OR a \code{SpatialExperiment} object.}

\item{...}{Additional arguments passed to methods.}

\item{meta_data}{Dataframe with metadata/covariates.
Ensure that rownames(meta_data) == colnames(x) (if x is a matrix).}

\item{sample_col}{String: Column name in meta_data identifying which rows correspond
to which sample (how to break up the data into multiple samples).}

\item{design_formula}{Formula object (compatible with stats::model.matrix).
Supply one of design_formula and design_mat.}

\item{design_mat}{Design matrix.
Supply one of design_formula and design_mat.
Zero columns will be dropped.}

\item{coord_data}{Dataframe or matrix with coordinates for observations.
Rows are samples. Ensure that rownames(coord_data) == colnames(x).
Needed to compute adjacency matrices.}

\item{D_THRESH}{Distance threshold for assigning two observations as adjacent.
Needed to compute adjacency matrices.}

\item{expected_num_neighbors}{In the absence of a known D_THRESH, how many
neighbors (on average) each cell should have. E.g., Visium data has 6.
Ignored if D_THRESH is supplied.}

\item{k_search}{When forming adjacency matrices, the maximum number of neighbors.}

\item{adj_mat}{A pre-computed adjacency matrix (sparse).}

\item{model_type}{One of "Leroux", "CAR", "SAR", "spNNGP", or "ALL".
Which model will be fit by TESSERA.
The value "ALL" will prepare the data for all four methods.}
}
\value{
A list comprised of the following:
\itemize{
\item \strong{coords_list}: A list with a matrix of coordinates for each sample.
\item \strong{covariates_list}: A list with a dataframe of metadata for each sample.
\item \strong{library_size_list}: A list with a vector of library sizes for each sample.
\item \strong{X_list}: A list with a matrix of design matrices for each sample.
\item \strong{W_list}: A list with measurement adjacency matrix for each sample.
\item \strong{D_list}: A list with measurement degree matrix for each sample.
\item \strong{eig_CS_list}: A list with CAR/SAR model eigenvalues for each sample.
\item \strong{eig_L_list}: A list with Leroux model eigenvalues for each sample.
}
}
\description{
Prepare data for the TESSERA method.
}
\note{
This function supports multiple dispatch for base matrices and SpatialExperiment objects.

SpatialExperiment support requires: "SpatialExperiment", "SingleCellExperiment", "SummarizedExperiment".
}
\examples{
# Locate the raw data in inst/extdata
rds_path <- system.file("extdata", "example_raw_data.rds", package = "TESSERA")
# Load the raw data list
raw_data <- readRDS(rds_path)

# Prepare data
TESSERA_data <- prepData(
  x = raw_data$count_matrix,
  meta_data = raw_data$meta_data,
  sample_col = "sample",
  design_mat = raw_data$design_mat,
  coord_data = raw_data$coords,
  adj_mat = raw_data$W,
  model_type = "Leroux"
)
}
\author{
Florica J Constantine, florica AT berkeley.edu
}
