% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepData.R
\name{prepData}
\alias{prepData}
\title{Prepare data for the TESSERA method.}
\usage{
prepData(
  count_matrix,
  meta_data,
  sample_col,
  design_formula = NULL,
  design_mat = NULL,
  coord_data = NULL,
  D_THRESH = NULL,
  k_search = 20,
  adj_mat = NULL,
  compute_eigs = "CSL"
)
}
\arguments{
\item{count_matrix}{Variables x samples (genes x cells) count matrix.
Can be sparse. Rownames are variable names.
Ensure that there is at least one row (a matrix).}

\item{meta_data}{Dataframe with metadata/covariates.
Ensure that rownames(meta_data) == colnames(count_matrix).}

\item{sample_col}{String: Column name in meta_data identifying which rows correspond
to which sample (how to break up the data into multiple samples).}

\item{design_formula}{Formula object (compatible with stats::model.matrix).
Supply one of design_formula and design_mat.}

\item{design_mat}{Design matrix.
Supply one of design_formula and design_mat.
Zero columns will be dropped.}

\item{coord_data}{Dataframe or matrix with coordinates for observations.
Rows are samples.
Ensure that rownames(coord_data) == colnames(count_matrix).
Needed to compute adjacency matrices.
Needed for spNNGP model, not necessary for lattice models if adjacency matrices
are supplied elsewhere.}

\item{D_THRESH}{Distance threshold for assigning two observations as adjacent.
Needed to compute adjacency matrices.
Not needed for non-lattice models (set to NULL).}

\item{k_search}{When forming adjacency matrices, the maximum number of neighbors.}

\item{adj_mat}{A pre-computed adjacency matrix (sparse).
Instead of forming it via D_THRESH or coordinates, a user can supply it.}

\item{compute_eigs}{Whether to compute eigenvalues of various matrices.
CAR/SAR: D^\{-1\} W; Leroux: D - W.
Depends on forming adjacency matrices, so needs coord_data, D_THRESH, k_search.
"None": Compute no eigenvalues.
String containing "C": Compute eigenvalues for the CAR model.
String containing "S": Compute eigenvalues for the SAR model.
String containing "L": Compute eigenvalues for the Leroux model.
If a non-lattice model (spNNGP) is desired, set this to "None"; otherwise,
match the lattice model desired.}
}
\value{
A list comprised of the following.

coords_list: A list with a matrix of coordinates for each sample.

covariates_list: A list with a dataframe of metadata for each sample.

library_size_list: A list with a vector of library sizes for each sample.
Total counts for each location if there are more than one gene/measurement;
otherwise, set to 1.

X_list: A list with a matrix of design matrices for each sample.

W_list: A list with measurement adjacency matrix for each sample.
If computed.

D_list: A list with measurement degree matrix for each sample.
If computed.

eig_CS_list: A list with CAR/SAR model eigenvalues for each sample.
If computed.

eig_L_list: A list with Leroux model eigenvalues for each sample.
If computed.
}
\description{
Prepare data for the TESSERA method.
}
\examples{
# Load package
library(TESSERA)

set.seed(2026)

## Generate synthetic data

data_list <- list()
n_samples <- 3 #
gamma_list <- c(0.1, 0.5, 0.9) # Spatial correlation parameter
tau2_list <- c(2.0, 0.1, 1.0) # Spatial variance parameter
beta_true <- c(1, 0, -1) # True fixed effects
n_list <- c(200, 300, 400) # Number of measurements per sample
nb_dist <- 0.05 # Distance to determine if a point is a neighbor
for (idx in seq_len(n_samples)) {
  data_list[[idx]] <- generate_data_one_area(
    n_points = n_list[idx],
    nb_dist = nb_dist,
    model_type = "Leroux",
    beta_true = beta_true,
    gamma_true = gamma_list[idx],
    tau2_true = tau2_list[idx]
  )
}

## Prepare for TESSERA

# This is how you'd go about preparing data manually
# See the prepDataSpatialExperiment for a much more streamlined workflow

# Count matrix: one gene per row
count_matrix <- matrix(Reduce(c, sapply(data_list, function (x) {
  x$z
})), nrow = 1)
rownames(count_matrix) <- c("example")
colnames(count_matrix) <- 1:ncol(count_matrix)
# Meta data
meta_data <- data.frame(sample = Reduce(c, sapply(seq_len(n_samples), function (x) {
  rep(x, n_list[x])
})))
rownames(meta_data) <- colnames(count_matrix)
# Coordinates
coords <- Reduce(rbind, sapply(data_list, function (x) {
  x$coords
}))
rownames(coords) <- colnames(count_matrix)
# Design matrix
design_mat <- Reduce(rbind, sapply(data_list, function (x) {
  x$X
}))
rownames(design_mat) <- colnames(count_matrix)
# Adjacency matrix
W <- Matrix::bdiag(sapply(data_list, function (x) {
  x$W
}))
rownames(W) <- colnames(count_matrix)
colnames(W) <- colnames(count_matrix)

# Prepare data
TESSERA_data <- prepData(
  count_matrix = count_matrix,
  meta_data = meta_data,
  sample_col = "sample",
  design_mat = design_mat,
  coord_data = coords,
  adj_mat = W,
  compute_eigs = "L"
)

# Optional checks for data objects
checkInputsTESSERA(TESSERA_data)
checkInputsTESSERAspNNGP(TESSERA_data)

## Fit models

# Note: In these examples we only run 2 iterations with 1 inner CM iteration.
# In practice, more iterations are needed.
suppressMessages(suppressWarnings({
  # Run TESSERA with a Leroux (lattice) model
  TESSERA_out_Leroux <- TESSERA_lattice(
    TESSERAData_obj = TESSERA_data,
    gene_name = "example",
    model_type = "Leroux",
    em_iters = 2,
    opt_iters = 1,
    verbose = FALSE
  )
  
  # Run TESSERA with a spNNGP (sparse Gaussian kernel) model
  TESSERA_out_spNNGP <- TESSERA_spNNGP(
    TESSERAData_obj = TESSERA_data,
    gene_name = "example",
    cov_type = "Mat",
    em_iters = 2,
    opt_iters = 1,
    verbose = FALSE
  )
}))

## Inference

# To get Wald statistics
waldTestStastics(TESSERAOutput_obj = TESSERA_out_Leroux,
                 contrast_mat = matrix(c(1, 0, 0), nrow = 1))

## Resampling / parametric data generation

# Resample the data using the same coordinates and covariates
# but with the fitted parameters
TESSERA_resampled_data <- prepSynthData(
  TESSERAData_obj = TESSERA_data,
  gene_list = "example",
  data_gen_model = "Leroux",
  tau2_true = TESSERA_out_Leroux$tau2_hat,
  gamma_true = TESSERA_out_Leroux$gamma_hat,
  beta_true = TESSERA_out_Leroux$beta_hat
)$new_TESSERAData_obj
# Refit on the synthetically generated data
# Note: In these examples we only run 2 iterations with 1 inner CM iteration.
# In practice, more iterations are needed.
suppressMessages(suppressWarnings({
  TESSERA_out_Leroux <- TESSERA_lattice(
    TESSERAData_obj = TESSERA_resampled_data,
    gene_name = "example",
    model_type = "Leroux",
    em_iters = 2,
    opt_iters = 1,
    verbose = FALSE
  )
}))
}
\author{
Florica J Constantine, florica AT berkeley.edu
}
