% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepData.R
\name{prepData}
\alias{prepData}
\title{Prepare data for the PoisECM method.}
\usage{
prepData(
  count_matrix,
  meta_data,
  sample_col,
  design_formula = NULL,
  design_mat = NULL,
  coord_data = NULL,
  D_THRESH = NULL,
  k_search = 20,
  adj_mat = NULL,
  compute_eigs = "CSL"
)
}
\arguments{
\item{count_matrix}{Variables x samples (genes x cells) count matrix.
Can be sparse. Rownames are variable names.
Ensure that there is at least one row (a matrix).}

\item{meta_data}{Dataframe with metadata/covariates.
Ensure that rownames(meta_data) == colnames(count_matrix).}

\item{sample_col}{String: Column name in meta_data identifying which rows correspond
to which sample (how to break up the data into multiple samples).}

\item{design_formula}{Formula object (compatible with stats::model.matrix).
Supply one of design_formula and design_mat.}

\item{design_mat}{Design matrix.
Supply one of design_formula and design_mat.
Zero columns will be dropped.}

\item{coord_data}{Dataframe or matrix with coordinates for observations.
Rows are samples.
Ensure that rownames(coord_data) == colnames(count_matrix).
Needed to compute adjacency matrices.
Needed for spNNGP model, not necessary for lattice models if adjacency matrices
are supplied elsewhere.}

\item{D_THRESH}{Distance threshold for assigning two observations as adjacent.
Needed to compute adjacency matrices.
Not needed for non-lattice models (set to NULL).}

\item{k_search}{When forming adjacency matrices, the maximum number of neighbors.}

\item{adj_mat}{A pre-computed adjacency matrix (sparse).
Instead of forming it via D_THRESH or coordinates, a user can supply it.}

\item{compute_eigs}{Whether to compute eigenvalues of various matrices.
CAR/SAR: D^{-1} W; Leroux: D - W.
Depends on forming adjacency matrices, so needs coord_data, D_THRESH, k_search.
"None": Compute no eigenvalues.
String containing "C": Compute eigenvalues for the CAR model.
String containing "S": Compute eigenvalues for the SAR model.
String containing "L": Compute eigenvalues for the Leroux model.
If a non-lattice model (spNNGP) is desired, set this to "None"; otherwise,
match the lattice model desired.}
}
\value{
A list comprised of the following.

coords_list: A list with a matrix of coordinates for each sample.

covariates_list: A list with a dataframe of metadata for each sample.

library_size_list: A list with a vector of library sizes for each sample.
Total counts for each location if there are more than one gene/measurement;
otherwise, set to 1.

X_list: A list with a matrix of design matrices for each sample.

W_list: A list with measurement adjacency matrix for each sample.
If computed.

D_list: A list with measurement degree matrix for each sample.
If computed.

eig_CS_list: A list with CAR/SAR model eigenvalues for each sample.
If computed.

eig_L_list: A list with Leroux model eigenvalues for each sample.
If computed.
}
\description{
Prepare data for the PoisECM method.
}
\author{
Florica J Constantine, florica AT berkeley.edu
}
