% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_gen.R
\name{prepSynthData}
\alias{prepSynthData}
\title{Prepare synthetic data for the PoisECM method.}
\usage{
prepSynthData(
  poisECMData_obj,
  gene_list,
  data_gen_model,
  tau2_true,
  gamma_true,
  beta_true
)
}
\arguments{
\item{poisECMData_obj}{Object containing data.
Created by the prepData method.}

\item{gene_list}{Which gene/measurement to fit.
I.e., Which row in the count data matrix to fit.
Rownames of \code{poisECMData_obj$counts_list[[idx]]}.
A string (single gene) or a vector of strings (multiple genes).}

\item{data_gen_model}{Which model to fit for the random effects.
"CAR", "SAR", and "Leroux" are the valid options.}

\item{tau2_true}{Spatial scale parameters.
A vector (single gene) or a matrix (genes x samples, rows are genes).}

\item{gamma_true}{Spatial correlation parameters.
A vector (single gene) or a matrix (genes x samples, rows are genes).}

\item{beta_true}{Coefficients.
A vector (single gene) or a matrix (genes x covariates, rows are genes).}
}
\value{
A list comprised of the following.

new_poisECMData_obj Object containing data.
Has all of the same information and values as poisECMData_obj,
except for new counts_list values.

synthetic_count_summary: A dataframe summarizing statistics of the
generated counts and computing various similarity statistics between the
generated and original counts.
}
\description{
Prepare synthetic data for the PoisECM method.
}
\author{
Florica J Constantine, florica AT berkeley.edu
}
