% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_gen.R
\name{prepSynthData}
\alias{prepSynthData}
\title{Prepare synthetic data for the PoisECM method.}
\usage{
prepSynthData(
  poisECMData_obj,
  gene_list,
  data_gen_model,
  tau2_true = NULL,
  gamma_true = NULL,
  cov_params = NULL,
  cov_type = NULL,
  nngp_k = NULL,
  beta_true = NULL
)
}
\arguments{
\item{poisECMData_obj}{Object containing data.
Created by the prepData method.}

\item{gene_list}{Which gene/measurement to fit.
I.e., Which row in the count data matrix to fit.
Rownames of \code{poisECMData_obj$counts_list[[idx]]}.
A string (single gene) or a vector of strings (multiple genes).}

\item{data_gen_model}{Which model to fit for the random effects.
"CAR", "SAR", and "Leroux" are the valid Lattice model options,
"spNNGP" is the only non-Lattice option.}

\item{tau2_true}{Spatial scale parameters.
A vector (single gene) or a matrix (genes x samples, rows are genes).
Only needed for lattice models.
Warning: 0 <= tau^2 is necessary.}

\item{gamma_true}{Spatial correlation parameters.
A vector (single gene) or a matrix (genes x samples, rows are genes).
Only needed for lattice models.
Warning: For CAR/SAR models, -1 < gamma < 1 is necessary.
Warning: For Leroux models, 0 <= gamma < 1 is necessary.
If these ranges are exceeded, errors are likely.}

\item{cov_params}{Spatial correlation parameters.
A matrix (samples x 3/4 parameters) or array (genes x samples x 3/4 parameters).
Only needed for spNNGP.
Parameters are (nugget, sill, range, and an optional smoothness (Matern)).}

\item{cov_type}{Spatial correlation kernel.
Only needed for spNNGP.
"Exp", "Mat", "Gau", and "Sph" are the valid options, for
Exponential, Matern, Gaussian, and Spherical, respectively.}

\item{nngp_k}{How many neighbors to use for the spNNGP kernel.}

\item{beta_true}{Coefficients.
A vector (single gene) or a matrix (genes x covariates, rows are genes).}
}
\value{
A list comprised of the following.

new_poisECMData_obj Object containing data.
Has all of the same information and values as poisECMData_obj,
except for new counts_list values.

synthetic_count_summary: A dataframe summarizing statistics of the
generated counts and computing various similarity statistics between the
generated and original counts.
}
\description{
Prepare synthetic data for the PoisECM method.
}
\author{
Florica J Constantine, florica AT berkeley.edu
}
