% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_initdata.R
\name{prune_graph}
\alias{prune_graph}
\title{Prunes mesh by eliminating long edges and small connected components}
\usage{
prune_graph(data, thresh_quantile = 0.95, mincells = 10, thresh = NA)
}
\arguments{
\item{data}{A list containing the mesh data structures:
\itemize{
\item \code{pts} is a V-by-M data table with columns \code{X} and \code{Y} containing the
coordinates of cells and additional metadata.
\item \code{tris} is a F-by-4 data table containing the X,Y coordinates of each
triangle's centroid in the first two columns, and area and
largest height of each triangle in the last two columns.
\item \code{edges} is a E-by-14 data table with columns \code{from_pt}, \code{to_pt}, \code{from_tri}, \code{to_tri},
\code{x0_pt}, \code{x1_pt}, \code{y0_pt}, \code{y1_pt}, \code{x0_tri}, \code{x1_tri}, \code{y0_tri}, \code{y1_tri},
\code{length_pt}, \code{length_tri}. If only one triangle uses an edge, then the \code{from_tri},
\code{x0_tri}, and \code{y0_tri} fields will contain NaN values.
\item \code{tri_to_pt} is a F-by-V sparse matrix with value 1 at (i,j) if
triangle i uses point j as a vertex.
}}

\item{thresh_quantile}{Floating point value between 0 and 1, inclusive.
Quantile of edge length above which edges are pruned. Defaults to 0.95.}

\item{mincells}{Minimum number of cells required for a connected
component of triangles to be kept. Defaults to 10.}

\item{thresh}{Edge length above which edges are pruned. If equal to NA,
then this value is ignored and \code{thresh_quantile} is used to compute
the threshold. Otherwise, if \code{thresh} is set, then \code{thresh_quantile}
is ignored. Defaults to NA.}
}
\value{
A list containing the mesh data structures with (possibly)
fewer points, edges, and triangles. Indices have been updated since
some objects might have been removed.
}
\description{
Pruning works in three steps:
\enumerate{
\item Any triangle with an edge longer than a threshold length is removed.
Afterwards, any edge that no longer belongs to a triangle is removed.
Then any point that no longer belongs to an edge is removed.
\item Connected components of triangles with a shared edge are computed.
If everything is connected, then nothing is pruned.
\item Otherwise, components (and corresponding triangles) that contain less
than \code{mincells} points are removed. Afterwards, any edge that no
longer belongs to a triangle is removed. Then any point that no longer
belongs to an edge is removed.
}
}
