% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_gen.R
\name{sample_Poisson_lattice}
\alias{sample_Poisson_lattice}
\title{Generate data from a Poisson lattice model with known parameters and adjacency matrix.}
\usage{
sample_Poisson_lattice(
  model_type,
  X,
  W,
  D,
  library_size,
  tau2_true,
  gamma_true,
  beta_true
)
}
\arguments{
\item{model_type}{"CAR", "SAR", or "Leroux"---model for random effects.}

\item{X}{Covariates.}

\item{W}{Adjacency matrix.}

\item{D}{Diagonal degree matrix (row-sums of W).}

\item{library_size}{Scaling for theta.}

\item{tau2_true}{True scale parameter.}

\item{gamma_true}{True correlation parameter.}

\item{beta_true}{True fixed effects.}
}
\value{
A list with the following fields:

z: Sampled counts Pois(theta x lib size).

phi_true: True sampled random effects.
(multivariate normal with mean zero and covariance tau^2 Q^\{-1\}).

eta_true: phi + X beta.

theta_true: exp(eta).
}
\description{
Generate data from a Poisson lattice model with known parameters and adjacency matrix.
}
\note{
Requires the Matrix library.

Calls the various Q_matrix creation functions.
}
\author{
Florica J Constantine, florica AT berkeley.edu
}
