% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_gen.R
\name{sample_Poisson_spNNGP}
\alias{sample_Poisson_spNNGP}
\title{Generate data from a Poisson spNNGP model with known parameters.}
\usage{
sample_Poisson_spNNGP(
  cov_type,
  X,
  library_size,
  coords,
  cov_params,
  nngp_k,
  beta_true
)
}
\arguments{
\item{cov_type}{Which model to fit for the Gaussian process covariance.
"Exp", "Mat", "Gau", and "Sph" are the valid options, for
Exponential, Matern, Gaussian, and Spherical, respectively.}

\item{X}{Covariates, rows are samples.}

\item{library_size}{Scaling for theta.}

\item{coords}{Matrix of coordinates.
Rows are samples, nrow(coords) == nrow(X).}

\item{cov_params}{Covariance parameters
(nugget, sill, range, smoothness (optional)).}

\item{nngp_k}{How many neighbors to use for the spNNGP kernel.}

\item{beta_true}{True fixed effects.}
}
\value{
A list with the following fields:

z: Sampled counts Pois(theta x lib size).

phi_true: True sampled random effects.
(multivariate normal with mean zero and covariance tau^2 Q^\{-1\}).

eta_true: phi + X beta.

theta_true: exp(eta).
}
\description{
Generate data from a Poisson spNNGP model with known parameters.
}
\note{
Requires the Matrix library.

Calls nngp_prec_mat.
}
\examples{
set.seed(2026)
cov_params_true <- c(1.0, 1.0, 10, 2)
beta_true <- c(1, 0, -1) 
ex_data <- generate_data_one_area_spNNGP(1000, 0.03, "Mat", cov_params_true, 20, beta_true)
sample_Poisson_spNNGP("Mat", ex_data$X, ex_data$library_size, cbind(ex_data$x_coords, ex_data$y_coords), cov_params_true, 20, beta_true)
}
\author{
Florica J Constantine, florica AT berkeley.edu
}
