% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/post_analysis.R
\name{selectWaldStatisticThreshold}
\alias{selectWaldStatisticThreshold}
\title{Given Wald statistics, compute an optimal threshold.}
\usage{
selectWaldStatisticThreshold(wald_stats, trimfrac = NULL)
}
\arguments{
\item{wald_stats}{Vector of Wald statistics: non-negative F-statistics.
Square t-statistics to obtain F-statistics.}

\item{trimfrac}{NULL if ignored, c(LOWER, UPPER) if not.
What fraction to trim from the left and right tails, e.g., c(0, 1e-4)
clips nothing from the left tail and the top 1e-4 fraction from the right tail,
i.e., everything above the 1 - 1e-4 quantile is trimmed.
In general, for this application, we recommend setting the lower/left value
to 0 and the right value to something small to drop outliers.}
}
\value{
Threshold value.
}
\description{
In the poisECM algorithm, and in generalized linear mixed models in general,
the exact distribution of p-values for Wald statistics under the null hypothesis
is unknown.
We know that the standard errors are, for finite samples, likely biased
downward (too small).
Moreover, for finite samples, due to the approximations
in the E-step of the fitting algorithm, there may be a small positive bias.
Hence, a scaled chi-squared distribution or scaled F-distribution with
a non-centrality parameter might be a good model for the statistics under
the null hypothesis, where we note that an unscaled chi-squared or F distribution
are the standard asymptotic distributions for Wald statistics.
To use this function, the Wald statistics as well as a threshold is estimated.
The threshold is a rough upper bound for the statistics coming from the null
distribution--it need not be exact, but it should be greater than most of the
null distributed statistics and smaller than most of the non-null distributed
statistics.
The threshold is estimated by noting that under the null hypothesis,
the p-values should be approximately Uniform(0, 1)
distributed; for each threshold value, we compute the mean-squared error
between the quantiles of the p-values and those of the Uniform(0, 1) distribution
and choose the threshold with the lowest mean-squared error.
}
\note{
Consider using the trimfrac parameter if the data have outliers:
This function is sensitive to the presence of large values/outliers.
}
\examples{
selectWaldStatisticThreshold(c(5 * stats::rchisq(2000, 1, ncp = 10),
 200 + 5 * stats::rchisq(2000, 1, ncp = 10)))
selectWaldStatisticThreshold(5 * stats::rchisq(2000, 1, ncp = 10))
}
\author{
Florica J Constantine, florica AT berkeley.edu
}
