% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tessera.R
\name{take.all.biopsies}
\alias{take.all.biopsies}
\title{Find the number of aneuploid cells}
\usage{
take.all.biopsies(embryo, n.cells.per.sample, chromosome)
}
\arguments{
\item{embryo}{an embryo as created by \code{create.embryo}}

\item{n.cells.per.sample}{the number of cells to take in each biopsy}

\item{chromosome}{the chromosome to test, or 0 for all chromosomes}
}
\value{
an integer vector of the number of aneuploid cells in each biopsy
}
\description{
Take all possible biopsies of the given size from the given
embryo
}
\examples{
e <- create.embryo(100, 0.1, 0.1)
take.all.biopsies(e, 5, 1)
}
