\name{tessera-package}
\alias{tessera}
\alias{tessera-package}
\docType{package}
\title{Genomic-driven biomarker identification for colorectal cancer}
\description{
  Statistics-focused R package supporting the in-silico validation of a
  colorectal-cancer polygenic risk score (CRC-PRS) against the circulating
  biomarker carcinoembryonic antigen (CEA) on a simulated UK-Biobank-like
  clinical cohort.
}
\keyword{internal}
