% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tessera-package.R
\docType{package}
\name{tessera}
\alias{tessera-package}
\alias{tessera}
\title{Accurate tiling of spatial single-cell data}
\description{
An algorithm for segmenting single-cell resolution
spatial omics data into small multicellular tiles whose edges track with
natural tissue boundaries. These tiles can then be used in downstream
analysis to label and define tissue regions across samples.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/korsunskylab/tessera}
  \item \url{https://korsunskylab.github.io/tessera/}
  \item Report bugs at \url{https://github.com/korsunskylab/tessera/issues}
}

}
\author{
\strong{Maintainer}: Ilya Korsunsky \email{ilya.korsunsky@gmail.com} (\href{https://orcid.org/0000-0003-4848-3948}{ORCID})

Authors:
\itemize{
  \item Daniel Stein \email{daniel_stein@g.harvard.edu} (\href{https://orcid.org/0000-0002-9442-2604}{ORCID})
}

}
