% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TESSERA_spNNGP.R
\name{TESSERA_spNNGP}
\alias{TESSERA_spNNGP}
\title{Fit multi-area Poisson spatial generalized linear model.
Allows for sparse nearest neighbor Gaussian process random effects.
Exponential, Gaussian, Matern, and Spherical covariance structures
are supported.
Fits a common set of fixed effects across areas.}
\usage{
TESSERA_spNNGP(
  TESSERAData_obj,
  gene_name,
  cov_type = "Exp",
  nngp_k = 20,
  em_iters = 200,
  opt_iters = 10,
  em_min_iters = 10,
  em_tol = 0.001,
  em_stopping = NULL,
  beta_init = "glm",
  cov_init = "BRISC",
  cov_fit_method = "BRISC",
  verbose = FALSE,
  dense_matrices = TRUE
)
}
\arguments{
\item{TESSERAData_obj}{Object containing data.
Created by the prepData method.}

\item{gene_name}{Which gene/measurement to fit.
I.e., Which row in the count data matrix to fit.}

\item{cov_type}{Which model to fit for the Gaussian process covariance.
"Exp", "Mat", "Gau", and "Sph" are the valid options, for
Exponential, Matern, Gaussian, and Spherical, respectively.}

\item{nngp_k}{How many neighbors to use for the spNNGP kernel.}

\item{em_iters}{Number of E(C)M iterations to run.}

\item{opt_iters}{Number of inner M-step iterations to run.
The M-step is technically a CM step, i.e., a conditional maximization.
That is, we perform block coordinate ascent wherein we maximize the
expected likelihood in tau^2 and then in gamma for each area, and then
in beta, where each maximization holds all other parameters constant.
This approach leads to a slower convergence in terms of the number of
iterations (em_iters) but faster computation in each iteration.}

\item{em_min_iters}{Minimum number of E(C)M iterations to run.}

\item{em_tol}{Tolerance for early stopping the E(C)M algorithm.}

\item{em_stopping}{How to decide to stop early.
NULL: Don't stop early.
"abs_loglike": Difference in absolute data log likelihood, |old - new|.
"rel_loglike": Relative difference in data log likelihood, |old - new|/|old|.
Data log likelihood taken across alll areas.
"abs_beta_norm": Absolute L2 norm of difference in beta, ||old - new||_2.
"rel_beta_norm": Relative L2 norm of difference in beta,
||old - new||_2 / ||old||_2.}

\item{beta_init}{How to initialize beta.
"glm": Stack up all the observed counts z and covariate matrices X and
fit a Poisson GLM z ~ 0 + X using base R's glm function.
"random": Initialize with standard normal random variables.
A vector or matrix: Initialize with provided value.}

\item{cov_init}{How to initialize covariance parameters.
"BRISC": Call BRISC on log(0.5 + z) - X beta_init.
"variogram": log(0.5 + z) - X beta_init.
A vector of matrix: Initialize with provided value.}

\item{cov_fit_method}{"BRISC" or "variogram": How to perform the M-step
estimation/update of covariance parameters.}

\item{verbose}{Boolean of whether to print out updates.}

\item{dense_matrices}{Boolean of whether to treat Q as a dense matrix when
computing certain quantities in the E-step. This will increase memory usage
(e.g., ~2 -> 8 GB for 3k cells per sample and 20 neighbors), but will dramatically
speed up computation (around a 60-80\% decrease).}
}
\value{
A list comprised of the following:

beta_hat: Estimated coefficients.

cov_param_hat: Estimated spatial covariance parameters.

predictions: Predicted values.
Vector of all values (corresponds to z_list stacked together).

eta_hat: Estimated random effects.
Vector of all values (corresponds to z_list stacked together).

theta_hat: Estimated Poisson parameters.
Vector of all values (corresponds to z_list stacked together).

phi_hat: Estimated spatial random effects.
Vector of all values (corresponds to z_list stacked together).

cov_param_tracker: Estimated correlation parameters.
areas x EM iterations x (Nugget, Sill, Range, Smoothness)---history across iterations.

beta_tracker: Estimated fixed effects.
coefficients x EM iterations matrix---history across iterations.

R2_tracker: Squared correlation between predictions and observations.
areas x EM Iterations matrix---history across iterations.

MSE_tracker: Mean-Square Error between predictions and observations.
areas x EM Iterations matrix---history across iterations.

data_log_like_tracker: Log likelihood of observations.
areas x EM Iterations matrix---history across iterations.

expected_log_like_tracker: Expected log likelihood at current parameters.
areas x EM Iterations matrix---history across iterations.

resid_moran_nb: Moran's I for each area computed using neighbor
adjacency as weights.
areas x EM Iterations matrix---history across iterations.

resid_moran: Moran's I for each area computed using coordinates.
areas x EM Iterations x (value, expectation, sd) array.
NULL unless coordinates are supplied.

start_idx_list: Indices where each area's values start in predictions, etc.
E.g., if area 1 has 100 points and area 2 has 50, we would have
c(1, 101, 151, ...).

time: Total time taken by function.

beta_neghessian: Negative Hessian of final value of beta.
Matrix.

run_settings: A list with the parameter settings used to run the algorithm.

performanceSummary: A dataframe with summary statistics for each sample.
See \code{\link[=summarizeTESSERAPerformance]{summarizeTESSERAPerformance()}} for more details.
}
\description{
Fit multi-area Poisson spatial generalized linear model.
Allows for sparse nearest neighbor Gaussian process random effects.
Exponential, Gaussian, Matern, and Spherical covariance structures
are supported.
Fits a common set of fixed effects across areas.
}
\note{
Requires the Matrix library.
}
\examples{
# Load package
library(TESSERA)

set.seed(2026)

## Generate synthetic data

data_list <- list()
n_samples <- 3 #
gamma_list <- c(0.1, 0.5, 0.9) # Spatial correlation parameter
tau2_list <- c(2.0, 0.1, 1.0) # Spatial variance parameter
beta_true <- c(1, 0, -1) # True fixed effects
n_list <- c(200, 300, 400) # Number of measurements per sample
nb_dist <- 0.05 # Distance to determine if a point is a neighbor
for (idx in seq_len(n_samples)) {
  data_list[[idx]] <- generate_data_one_area(
    n_points = n_list[idx],
    nb_dist = nb_dist,
    model_type = "Leroux",
    beta_true = beta_true,
    gamma_true = gamma_list[idx],
    tau2_true = tau2_list[idx]
  )
}

## Prepare for TESSERA

# This is how you'd go about preparing data manually
# See the prepDataSpatialExperiment for a much more streamlined workflow

# Count matrix: one gene per row
count_matrix <- matrix(Reduce(c, sapply(data_list, function (x) {
  x$z
})), nrow = 1)
rownames(count_matrix) <- c("example")
colnames(count_matrix) <- 1:ncol(count_matrix)
# Meta data
meta_data <- data.frame(sample = Reduce(c, sapply(seq_len(n_samples), function (x) {
  rep(x, n_list[x])
})))
rownames(meta_data) <- colnames(count_matrix)
# Coordinates
coords <- Reduce(rbind, sapply(data_list, function (x) {
  x$coords
}))
rownames(coords) <- colnames(count_matrix)
# Design matrix
design_mat <- Reduce(rbind, sapply(data_list, function (x) {
  x$X
}))
rownames(design_mat) <- colnames(count_matrix)
# Adjacency matrix
W <- Matrix::bdiag(sapply(data_list, function (x) {
  x$W
}))
rownames(W) <- colnames(count_matrix)
colnames(W) <- colnames(count_matrix)

# Prepare data
TESSERA_data <- prepData(
  count_matrix = count_matrix,
  meta_data = meta_data,
  sample_col = "sample",
  design_mat = design_mat,
  coord_data = coords,
  adj_mat = W,
  compute_eigs = "L"
)

# Optional checks for data objects
checkInputsTESSERA(TESSERA_data)
checkInputsTESSERAspNNGP(TESSERA_data)

## Fit models

# Note: In these examples we only run 2 iterations with 1 inner CM iteration.
# In practice, more iterations are needed.
suppressMessages(suppressWarnings({
  # Run TESSERA with a Leroux (lattice) model
  TESSERA_out_Leroux <- TESSERA_lattice(
    TESSERAData_obj = TESSERA_data,
    gene_name = "example",
    model_type = "Leroux",
    em_iters = 2,
    opt_iters = 1,
    verbose = FALSE
  )
  
  # Run TESSERA with a spNNGP (sparse Gaussian kernel) model
  TESSERA_out_spNNGP <- TESSERA_spNNGP(
    TESSERAData_obj = TESSERA_data,
    gene_name = "example",
    cov_type = "Mat",
    em_iters = 2,
    opt_iters = 1,
    verbose = FALSE
  )
}))

## Inference

# To get Wald statistics
waldTestStastics(TESSERAOutput_obj = TESSERA_out_Leroux,
                 contrast_mat = matrix(c(1, 0, 0), nrow = 1))

## Resampling / parametric data generation

# Resample the data using the same coordinates and covariates
# but with the fitted parameters
TESSERA_resampled_data <- prepSynthData(
  TESSERAData_obj = TESSERA_data,
  gene_list = "example",
  data_gen_model = "Leroux",
  tau2_true = TESSERA_out_Leroux$tau2_hat,
  gamma_true = TESSERA_out_Leroux$gamma_hat,
  beta_true = TESSERA_out_Leroux$beta_hat
)$new_TESSERAData_obj
# Refit on the synthetically generated data
# Note: In these examples we only run 2 iterations with 1 inner CM iteration.
# In practice, more iterations are needed.
suppressMessages(suppressWarnings({
  TESSERA_out_Leroux <- TESSERA_lattice(
    TESSERAData_obj = TESSERA_resampled_data,
    gene_name = "example",
    model_type = "Leroux",
    em_iters = 2,
    opt_iters = 1,
    verbose = FALSE
  )
}))
}
\references{
Meng, Xiao-Li, and Donald B. Rubin.
"Maximum likelihood estimation via the ECM algorithm: A general framework."
Biometrika 80.2 (1993): 267-278.
}
\author{
Florica J Constantine, florica AT berkeley.edu
}
