% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepData.R
\name{visualizeNeighborDistances}
\alias{visualizeNeighborDistances}
\title{Prepare data for the TESSERA method: Choose a distance threshold for
adjacency matrices.}
\usage{
visualizeNeighborDistances(meta_data, sample_col, coord_data, k_search = 20)
}
\arguments{
\item{meta_data}{Dataframe with metadata/covariates.}

\item{sample_col}{String: Column name in meta_data identifying which rows correspond
to which sample.}

\item{coord_data}{Dataframe or matrix with coordinates for observations.}

\item{k_search}{When forming adjacency matrices, the maximum number of neighbors.}
}
\value{
A list comprised of distances and a ggplot2 visualization.
}
\description{
Prepare data for the TESSERA method: Choose a distance threshold for
adjacency matrices.
}
\author{
Florica J Constantine, florica AT berkeley.edu
}
