% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/post_analysis.R
\name{waldStatisticPValuesThreshold}
\alias{waldStatisticPValuesThreshold}
\title{Given Wald statistics, compute p-values.}
\usage{
waldStatisticPValuesThreshold(wald_stats, threshold)
}
\arguments{
\item{wald_stats}{Vector of Wald statistics: non-negative F-statistics.
Square t-statistics to obtain F-statistics.}

\item{threshold}{A rough guess or upper bound for the Wald statistics under
the null hypothesis.}
}
\value{
Vector of p-values (un-adjusted for multiple corrections).
}
\description{
In the poisECM algorithm, and in generalized linear mixed models in general,
the exact distribution of p-values for Wald statistics under the null hypothesis
is unknown.
We know that the standard errors are, for finite samples, likely biased
downward (too small).
Moreover, for finite samples, due to the approximations
in the E-step of the fitting algorithm, there may be a small positive bias.
Hence, a scaled chi-squared distribution or scaled F-distribution with
a non-centrality parameter might be a good model for the statistics under
the null hypothesis, where we note that an unscaled chi-squared or F distribution
are the standard asymptotic distributions for Wald statistics.
To use this function, the Wald statistics as well as a threshold are required.
The threshold is a rough upper bound for the statistics coming from the null
distribution--it need not be exact, but it should be greater than most of the
null distributed statistics and smaller than most of the non-null distributed
statistics.
In the absence of oracle knowledge, we recommend the \code{selectWaldStatisticThreshold}
function in this package for choosing an optimal threshold value.
}
\examples{
waldStatisticPValuesThreshold(c(
  5 * stats::rchisq(2000, 1, ncp = 10),
  200 + 5 * stats::rchisq(2000, 1, ncp = 10)
 ), 200)
}
\author{
Florica J Constantine, florica AT berkeley.edu
}
