% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/post_analysis.R
\name{waldTestStastics}
\alias{waldTestStastics}
\title{Given the output of the TESSERA algorithms and a contrast matrix, compute
Wald T-statistics.}
\usage{
waldTestStastics(TESSERAOutput_obj, contrast_mat)
}
\arguments{
\item{TESSERAOutput_obj}{Output of the TESSERA algorithms.}

\item{contrast_mat}{Matrix with contrasts of the estimated coeficients.
Rows are contrasts, columns correspond to columns in beta_hat (covariates).
beta_hat is the vector of estimated coefficients, stored in TESSERAOutput_obj.}
}
\value{
A dataframe of Wald t-statistics. Square these to get F-statistics.
}
\description{
Given the output of the TESSERA algorithms and a contrast matrix, compute
Wald T-statistics.
}
\examples{
# Locate the saved model results in inst/extdata
rds_path <- system.file("extdata", "example_TESSERA_out_Leroux.rds", package = "TESSERA")
# Load the results object
TESSERA_out_Leroux <- readRDS(rds_path)

# Run the Wald test
waldTestStastics(TESSERAOutput_obj = TESSERA_out_Leroux,
                 contrast_mat = matrix(c(1, 0, 0), nrow = 1))
}
\author{
Florica J Constantine, florica AT berkeley.edu
}
