% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/post_analysis.R
\name{waldTestStastics}
\alias{waldTestStastics}
\title{Given the output of the poisECM algorithms and a contrast matrix, compute
Wald T-statistics.}
\usage{
waldTestStastics(poisECMDOutput_obj, contrast_mat)
}
\arguments{
\item{poisECMDOutput_obj}{Output of the poisECM algorithms.}

\item{contrast_mat}{Matrix with contrasts of the estimated coeficients.
Rows are contrasts, columns correspond to columns in beta_hat (covariates).}
}
\value{
A dataframe of Wald t-statistics.
}
\description{
Given the output of the poisECM algorithms and a contrast matrix, compute
Wald T-statistics.
}
